==> _p_ADAPTOmics/_I_Desiree/_S_multiOmics/_A_multiOmicsStat-R/scripts/01_DEP-prot.Rmd <==
---
title: "01_DEP-prot"
author: "zagor"
date: "`r Sys.Date()`"
output:
  html_document:
    fig_caption: yes
    self_contained: yes
    fig_width: 9
    fig_height: 8
    toc: yes
    toc_float: true
    code_folding: show
    toc_depth: 5
    number_sections: yes
    theme: flatly
    highlight: tango
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE, echo=FALSE, warning=FALSE, tidy=TRUE}
knitr::opts_chunk$set(#dev = c('pdf', 'png'),  # this embeds pdf and crates scrolable blocks
                      dev = c('png')
                      , fig.align = 'center'
                      , fig.height = 8
                      , fig.width = 9 
                      , warning = FALSE
                      , message = FALSE
                      , cache.extra = TRUE)
                      
# options(knitr.table.format = "html")


```



```{r, echo=TRUE, warning=FALSE, cache.extra = TRUE}

rm(list = ls(all = TRUE))
gc()

seed = 123456
set.seed(seed)


```

# libraries

```{r, libraries, echo=TRUE, warning=FALSE, cache.extra = TRUE}

# if (!require("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")

# current version
# BiocManager::install(version = "3.20")

# manuscript version
## To use Bioconductor version '3.17', first downgrade XY packages with
## BiocManager::install(version = '3.17')
# BiocManager::install("DEP")


# can be useful
# BiocManager::install("fgsea")
# BiocManager::install("GEOquery")
# BiocManager::install("limma", force = TRUE)
# BiocManager::install("BiocGenerics")
# BiocManager::install("mzR", force = TRUE)
# devtools::install_github("mildpiggy/DEP2", force = TRUE)



library(magrittr)
library(DEP)



```

# define comparisons/contrasts

or write them manually

```{r, comparisons, echo=TRUE, warning=FALSE}

fp = file.path('..', 'input')
fn = 'forComparisons.txt'

forComparisons = data.table::fread(file.path(fp, fn), header = TRUE)
data.table::setDF(forComparisons)

selection = cbind(c(forComparisons$stress, forComparisons$control), 
                  c(forComparisons$timeS, forComparisons$timeC))
selection = as.data.frame(selection[!duplicated(selection), ])
colnames(selection) = c('Treatment', 'SamplingDay')
selection$SamplingDay = as.numeric(selection$SamplingDay)
selection$TreatmentxSamplingDay = paste(selection$Treatment, selection$SamplingDay, sep = '_')

selection = selection[order(selection$SamplingDay, selection$Treatment), ]
# print(selection)


```


## contrasts


| **treatment** | **time point** | vs | **treatment** | **time point**|
|:------------|:------------|:------:|------------:|------------:|
| $\color{DarkRed}{\text{H}}$ | 1 | - | C | 1 |
| $\color{blue}{\text{W}}$ | 1 | - | C | 1 |
| $\color{DarkRed}{\text{H}}$ | 7 | - | C | 7 |
| $\color{blue}{\text{W}}$ | 7 | - | C | 7 |
| $\color{DarkRed}{\text{H}}$ | 8 | - | C | 8 |
| $\color{orange}{\text{D}}$ | 8 | - | C | 8 |
| $\color{OrangeRed}{\text{HD}}$ | 8 | - | C | 8 |
| $\color{DarkRed}{\text{H}}$ | 14 | - | C | 14 |
| $\color{orange}{\text{D}}$ | 14 | - | C | 14 |
| $\color{OrangeRed}{\text{HD}}$ | 14 | - | C | 14 |
| $\color{DarkRed}{\text{H}}$ | 8 | - | $\color{OrangeRed}{\text{HD}}$ | 8 |
| $\color{DarkRed}{\text{H}}$ | 14 | - | $\color{OrangeRed}{\text{HD}}$ | 14 |
...



# read proteomics measurements

```{r, data, echo=TRUE, warning=FALSE}

fn = 'data_Proteomics.xlsx'
fl = openxlsx::getSheetNames(file.path(fp, fn))

cnt = TRUE
Prot = NULL

for (i in fl) {
  
  print(i)
  
  tmp = as.data.frame(openxlsx::read.xlsx(file.path(fp, fn), sheet = i))
  colnames(tmp)[1] = 'pID'
  tmp$pID = gsub(' ', '', tmp$pID)
  rownames(tmp) = tmp$pID 

  if(cnt){
    Prot = tmp
    cnt = FALSE
  } else {
    Prot = merge(Prot, tmp, by = 'pID', all.x = TRUE, all.y = TRUE)
  }
  
}


```


```{r, NAs, echo=TRUE, warning=FALSE}


# NA and 0

# table(is.na(Prot))
# table(Prot == 0)
Prot[Prot == 0] = NA



rownames(Prot) = Prot$pID
myProteomics = as.matrix(Prot[, -grep('pID', colnames(Prot))])
# typeof(myProteomics)


tmp = as.data.frame(do.call(rbind, stringr::str_split(colnames(myProteomics), '_')))
tmp = paste(tmp[, 1], tmp[, 2], sep = '_')
tmp = gsub('S', '', tmp)
myProteomics = myProteomics[, tmp %in% selection$TreatmentxSamplingDay]

cn = colnames(myProteomics)
rn = rownames(myProteomics)
# typeof(myProteomics)
# range(myProteomics, na.rm = TRUE)


myProteomics0 = myProteomics
myProteomics0[is.na(myProteomics0)] = 0


```

# density

```{r, density, echo=TRUE, warning=FALSE}



plot(density(myProteomics0),  main = 'NA as 0; "raw"')
plot(density(log(myProteomics0, 10)), main = 'NA as 0; log10')
plot(density(log(myProteomics0, 2)), main = 'NA as 0; log2')

rm(myProteomics0)


P.l = tidyr::gather(Prot, group, value, colnames(Prot)[2]:colnames(Prot)[ncol(Prot)], factor_key = FALSE)
# head(P.l)
tmp = as.data.frame(do.call(rbind, stringr::str_split(P.l$group, '_')))
tmp[, 2] = stringr::str_pad(gsub('S', '', tmp[, 2]), 2, pad = "0")
tmp = paste(tmp[, 1], tmp[, 2], sep = '_')


P.l$group = tmp
P.l = P.l[P.l$group %in% paste0(gsub('_.*', '_', selection$TreatmentxSamplingDay),
                                stringr::str_pad(gsub('.*_', '', selection$TreatmentxSamplingDay), 2, pad = "0")), ]
P.l$group = as.factor(P.l$group)

ggplot2::ggplot(data = P.l, ggplot2::aes(x = log(value, 2), group = group, fill = group)) +
    ggplot2::geom_density(adjust=1.5, alpha=.4)

rm(P.l)

```


# gmt

```{r, gmt, echo=TRUE, warning=FALSE}

fn = 'stu_StPanTr_3_CDS_2021-02-26_mapping_manual.gmt'
tmp = (readLines(file.path(fp, fn)))
# length(tmp)
# typeof(tmp)
# typeof(tmp[1])
n = stringr::str_count(tmp, pattern = '\t')
l = tmp[which(n == max(n))]
m = max(n+1)

temp = stringr::str_split_fixed(tmp, '\t', m)
# dim(temp)
# typeof(temp)
gmm = as.data.frame(temp)

pathwaysDF = tidyr::gather(gmm, column, proteinID, colnames(gmm)[2]:colnames(gmm)[ncol(gmm)], factor_key = FALSE)
pathwaysDF = pathwaysDF[, -2]
colnames(pathwaysDF) = c('bin', 'pID')
# head(pathwaysDF)
# table(pathwaysDF$ID == '')
# table(is.na(pathwaysDF$ID == ''))
pathwaysDF = pathwaysDF[pathwaysDF$pID != '', ]

rm(gmm)
gc()



```




# DEP

<https://bioconductor.org/packages/release/bioc/vignettes/DEP/inst/doc/DEP.html>


- log2 values

- no additional normalisation

- no imputation (when needed see <https://bioconductor.org/packages/release/bioc/vignettes/DEP/inst/doc/MissingValues.html#data-imputation-of-missing-data>)

- Protein-wise linear models combined with empirical Bayes statistics are used for the differential enrichment analysis (or differential expression analysis)

- The test_diff function introduced here uses limma and automatically generates the contrasts to be tested

- For the contrasts generation, the control sample has to be specified. Additionally, the types of contrasts to be produced need to be indicated, allowing the generation of all possible comparisons (“all”) or the generation of contrasts of every sample versus control (“control”). Alternatively, the user can manually specify the contrasts to be tested (type = “manual”), which need to be specified in the argument test.



```{r, DEP, echo=TRUE, warning=FALSE}


# table(is.na(myProteomics))
# table(myProteomics == 0)
# 
# colnames(myProteomics)

LFQ_columns = 1:ncol(myProteomics)
data_unique = myProteomics
experimental_design = as.data.frame(colnames(myProteomics))
colnames(experimental_design) = 'label'
tmp = (strsplit(colnames(myProteomics), '_'))
tmp = as.data.frame(do.call(rbind, tmp))  
tmp = paste(tmp[, 1], tmp[, 2], sep = '_')
experimental_design$condition = tmp
experimental_design$replicate = rep(1:4, times = ncol(myProteomics)/4)

df = as.data.frame(myProteomics)
df$Protein.IDs = rownames(myProteomics)
df$Protein.names = rownames(myProteomics)
data_unique = make_unique(proteins = df, ids = "Protein.IDs", names = "Protein.names")

rm(myProteomics)

```


## protein identification

```{r, identification, echo=TRUE, warning=FALSE}



# Are there any duplicated names?
# data_unique$name %>% duplicated() %>% any()


# Generate a SummarizedExperiment object using an experimental design
data_se = make_se(data_unique, LFQ_columns, experimental_design)
# data_se
# all(paste(data_se@assays@data[[1]]) == paste(as.matrix(log2(data_unique[, 1:56]))))
rm(data_unique)

plot_frequency(data_se)

```


## filtering and normalisation

```{r, filtering, echo=TRUE, warning=FALSE}


# filter_missval filters a proteomics dataset based on missing values. 
## The dataset is filtered for proteins that have a maximum of 'thr' missing values in at least one condition.
data_filt = filter_missval(data_se, thr = 0)
rm(data_se)
# data_filt

# Plot a barplot of the number of identified proteins per samples
plot_numbers(data_filt)


# Plot a barplot of the protein identification overlap between samples
plot_coverage(data_filt)



```



```{r, missval, echo=FALSE, warning=FALSE}

plot_missval = function (se) {
    assertthat::assert_that(inherits(se, "SummarizedExperiment"))
    se_assay <- SummarizedExperiment::assay(se)
    
    if (!any(is.na(se_assay))) {
        stop("No missing values in '", deparse(substitute(se)), 
            "'", call. = FALSE)
    }
    
    df <- se_assay %>% data.frame(.)
    missval <- df[apply(df, 1, function(x) any(is.na(x))), ]
    missval <- ifelse(is.na(missval), 0, 1)
    
    ht2 = ComplexHeatmap::Heatmap(missval, 
                  col = c("grey60", "navy"), 
        column_names_side = "top", 
        show_row_names = FALSE, 
        show_column_names = TRUE, 
        name = "Missing values pattern", 
        column_names_gp = grid::gpar(fontsize = 5), 
        heatmap_legend_param = list(at = c(0, 1), labels = c("Missing value", "Valid value")))
    
    ComplexHeatmap::draw(ht2, heatmap_legend_side = "top")
}


```



```{r, plotDensities, echo=TRUE, warning=FALSE}


plot_missval(data_filt)
plot_detect(data_filt)

plot_normalization(data_filt)
# no additional normalisation

# quantile
data_quantile = data_filt
data_quantile@assays@data[[1]] = as.data.frame(preprocessCore::normalize.quantiles(as.matrix(data_quantile@assays@data[[1]])))
plot_normalization(data_quantile)

data_VSN = data_filt
data_VSN@assays@data[[1]] = as.data.frame(limma::normalizeVSN(as.matrix(data_VSN@assays@data[[1]])))
plot_normalization(data_VSN)



data_norm = data_filt
# data_norm@assays@data[[1]]) is filtered log(myProteomics, 2)
limma::plotDensities(data_norm@assays@data[[1]], main = 'Filtered, log2"', legend = FALSE)
limma::plotDensities(data_quantile@assays@data[[1]], main = 'Quantile', legend = FALSE)
limma::plotDensities(data_VSN@assays@data[[1]], main = 'VSN', legend = FALSE)


rm(data_filt)
rm(data_quantile)
rm(data_VSN)


```


```{r, threshold, echo=TRUE, warning=FALSE}

threshold = 0.05
lfc.threshold = 0


```


## PCA

```{r, PCA, echo=TRUE, warning=FALSE}

set.seed(seed)

myContrasts = paste(paste(forComparisons$stress, forComparisons$timeS, sep = '_S'),
                    paste(forComparisons$control, forComparisons$timeC, sep = '_S'),
                    sep = '_vs_')

# Test manually defined comparisons
data_diff_manual = test_diff(data_norm, 
                             type = "manual", 
                             test = myContrasts)


# Denote significant proteins based on user defined cutoffs
dep = add_rejections(data_diff_manual, alpha = threshold, lfc = lfc.threshold)

rs = rowSums(data_norm@assays@data[[1]])
# table(is.na(rs))
m = sum(!is.na(rs))

plot_pca(dep, x = 1, y = 2, point_size = 4, indicate = "condition", label = TRUE, n = m) # n = nrow(dep)
plot_pca(dep, x = 1, y = 3, point_size = 4, indicate = "condition", label = TRUE, n = m) # n = nrow(dep)


```


## Gower’s distance 

computed as the average of partial dissimilarities across individuals. 

```{r, Gower, echo=TRUE, warning=FALSE}

plot_dist(dep, significant = TRUE, pal = "Spectral", font_size = 5)


```


## heatmap

of all significant proteins with the data centered per protein

```{r, Heatmap, echo=FALSE, warning=FALSE}




plot_heatmapX = function (dep, type = c("contrast", "centered"), 
    kmeans = FALSE, k = 6, col_limit = 6, indicate = NULL, clustering_distance = c("euclidean", 
        "maximum", "manhattan", "canberra", 
        "binary", "minkowski", "pearson", "spearman", 
        "kendall", "gower"), row_font_size = 6, col_font_size = 10, 
    plot = TRUE, ...) 
{
    if (is.integer(k)) 
        k <- as.numeric(k)
    if (is.integer(col_limit)) 
        col_limit <- as.numeric(col_limit)
    if (is.integer(row_font_size)) 
        row_font_size <- as.numeric(row_font_size)
    if (is.integer(col_font_size)) 
        col_font_size <- as.numeric(col_font_size)
    
    assertthat::assert_that(inherits(dep, "SummarizedExperiment"), 
        is.character(type), is.logical(kmeans), is.numeric(k), 
        length(k) == 1, is.numeric(col_limit), length(col_limit) == 
            1, is.numeric(row_font_size), length(row_font_size) == 
            1, is.numeric(col_font_size), length(col_font_size) == 
            1, is.logical(plot), length(plot) == 1)
    
    type <- match.arg(type)
    clustering_distance <- match.arg(clustering_distance)
    
    row_data <- SummarizedExperiment::rowData(dep, use.names = FALSE)
    col_data <- SummarizedExperiment::colData(dep) %>% as.data.frame()
    
    if (any(!c("label", "condition", "replicate") %in% 
        colnames(col_data))) {
        stop(paste0("'label', 'condition' and/or 'replicate' columns are not present in '", 
            deparse(substitute(dep)), "'"), call. = FALSE)
    }
    
    if (length(grep("_diff", colnames(row_data))) < 1) {
        stop(paste0("'[contrast]_diff' columns are not present in '", 
            deparse(substitute(dep)), "'.\nRun test_diff() to obtain the required columns."), 
            call. = FALSE)
    }
    
    if (!"significant" %in% colnames(row_data)) {
        stop(paste0("'significant' column is not present in '", 
            deparse(substitute(dep)), "'.\nRun add_rejections() to obtain the required column."), 
            call. = FALSE)
    }
    
    if (!is.null(indicate) & type == "contrast") {
        warning("Heatmap annotation only applicable for type = 'centered'", 
            call. = FALSE)
    }
    
    if (!is.null(indicate) & type == "centered") {
        ha1 <- get_annotation(dep, indicate)
    } else {
        ha1 <- NULL
    }
    
    filtered <- dep[row_data$significant, ]
    
    if (any(is.na(SummarizedExperiment::assay(filtered)))) {
        warning("Missing values in '", deparse(substitute(dep)), 
            "'. ", "Using clustering_distance = 'gower'", 
            call. = FALSE)
        clustering_distance <- "gower"
        obs_NA <- TRUE
    } else {
        obs_NA <- FALSE
    }
    
    if (type == "centered") {
        rowData(filtered)$mean <- rowMeans(SummarizedExperiment::assay(filtered), na.rm = TRUE)
        df <- assay(filtered) - rowData(filtered, use.names = FALSE)$mean
    }
    
    if (type == "contrast") {
        df <- SummarizedExperiment::rowData(filtered, use.names = FALSE) %>% data.frame() %>% 
            tibble::column_to_rownames(var = "name") %>% dplyr::select(ends_with("_diff"))
        colnames(df) <- gsub("_diff", "", colnames(df)) %>% 
            gsub("_vs_", " vs ", .)
        df <- as.matrix(df)
    }
    
    if (kmeans & obs_NA) {
        warning("Cannot perform kmeans clustering with missing values", 
            call. = FALSE)
        kmeans <- FALSE
    }
    
    if (kmeans & !obs_NA) {
        set.seed(seed)
        df_kmeans <- kmeans(df, k)
        if (type == "centered") {
            order <- data.frame(df) %>% cbind(., cluster = df_kmeans$cluster) %>% 
                mutate(row = apply(.[, seq_len(ncol(.) - 1)], 
                  1, function(x) max(x))) %>% group_by(cluster) %>% 
                summarize(index = sum(row)/n()) %>% arrange(desc(index)) %>% 
                pull(cluster) %>% match(seq_len(k), .)
            df_kmeans$cluster <- order[df_kmeans$cluster]
        }
        if (type == "contrast") {
            order <- data.frame(df) %>% cbind(df, cluster = df_kmeans$cluster) %>% 
                gather(condition, diff, -cluster) %>% group_by(cluster) %>% 
                summarize(row = mean(diff)) %>% arrange(desc(row)) %>% 
                pull(cluster) %>% match(seq_len(k), .)
            df_kmeans$cluster <- order[df_kmeans$cluster]
        }
    }
    
    if (ncol(df) == 1) {
        col_clust = FALSE
    } else {
        col_clust = TRUE
    }
    
    if (nrow(df) == 1) {
        row_clust = FALSE
    } else {
        row_clust = FALSE # 2023-04-25
    }
    
    if (clustering_distance == "gower") {
        clustering_distance <- function(x) {
            dist <- cluster::daisy(x, metric = "gower")
            dist[is.na(dist)] <- max(dist, na.rm = TRUE)
            return(dist)
        }
    }

    myorder = c("H_S1 vs C_S1", 
                "W_S1 vs C_S1",

                "H_S7 vs C_S7", 
                "W_S7 vs C_S7",
                
                "D_S8 vs C_S8", 
                "HD_S8 vs C_S8", 
                "H_S8 vs C_S8",
                
                "D_S14 vs C_S14",
                "HD_S14 vs C_S14",
                "H_S14 vs C_S14"
                )
    
    ind = match(myorder, colnames(df))
    df = df[, myorder]
    colnames(df) = gsub('_S', '', colnames(df))
    
    legend <- ifelse(type == "contrast", "log2 Fold change", 
        "log2 Centered intensity")
    ht1 = ComplexHeatmap::Heatmap(df, col = circlize::colorRamp2(seq(-col_limit, 
        col_limit, (col_limit/5)), rev(RColorBrewer::brewer.pal(11, 
        "RdBu"))), split = if (kmeans) {
        df_kmeans$cluster
    } else {
        NULL
    }, cluster_rows = col_clust, cluster_columns = row_clust, 
        row_names_side = "left", column_names_side = "top", 
        clustering_distance_rows = clustering_distance, clustering_distance_columns = clustering_distance, 
        heatmap_legend_param = list(color_bar = "continuous", 
            legend_direction = "horizontal", legend_width = grid::unit(5, 
                "cm"), title_position = "lefttop"), 
        name = legend, row_names_gp = grid::gpar(fontsize = row_font_size), 
        column_names_gp = grid::gpar(fontsize = col_font_size), top_annotation = ha1, 
    # https://github.com/jokergoo/ComplexHeatmap/issues/370
    column_split = c(1, 1, 2, 2, 3,3,3, 4,4,4), column_title = NULL, # 2023-04-25
        ...)
    
    if (plot) {
        ComplexHeatmap::draw(ht1, heatmap_legend_side = "top")
    } else {
        colnames(df) <- gsub(" ", "_", colnames(df))
        df <- df[, unlist(column_order(ht1))]
        if (kmeans) {
            df <- cbind(df, k = df_kmeans$cluster)
        }
        return <- df[unlist(row_order(ht1)), ]
        data.frame(protein = row.names(return), return) %>% mutate(order = row_number())
    }
    
}

```



```{r, heatmapX, echo=TRUE, warning=FALSE}

# Plot a heatmap of all significant proteins with the data centered per protein


# range(dep@assays@data[[1]], na.rm = TRUE)
# dim(dep@assays@data[[1]])

plot_heatmapX(dep = dep, type = 'contrast', col_limit = 3.0, row_font_size = 3,
             clustering_distance = 'gower',
             col_font_size = 8)


```

```{r, good-to-know-section, echo=FALSE, warning=FALSE, results='hide'}

# good to know section
# https://support.bioconductor.org/p/p133303/
# tmp = matrix(unlist(dep@elementMetadata@listData), ncol=ncol(dep@elementMetadata))
# colnames(tmp) = names(dep@elementMetadata@listData)
# p = as.matrix(tmp[, grep('p\\.val', colnames(tmp))])
# ind = grep('p\\.adj', colnames(tmp))
# p.adj = as.matrix(tmp[, ind])
# all(gsub('_p.*', '', colnames(p)) == gsub('_p.*', '', colnames(p.adj)))
# 
# 
# apply(p.adj, 2, function(x) table(as.numeric(x) < threshold))
# 
# p.adj.BH = apply(p, 2, function(x) p.adjust(as.numeric(x), method = 'fdr'))
# apply(p.adj.BH, 2, function(x) table(as.numeric(x) < threshold))




```



## volcano plots



```{r, Volcano, echo=TRUE, warning=FALSE}

plots = list()
cnt = 1
for (i in setdiff(myContrasts, c("D_S8_vs_C_S8", "D_S14_vs_C_S14" ))) {
  contrast = NULL
  assign('contrast', value = i)
  p = plot_volcano(dep, contrast = contrast, label_size = 2, add_names = TRUE, adjusted = FALSE) 
  plots[[cnt]] = p
  cnt = cnt + 1
}

print(plots)


```


## frequency


of significant conditions per protein and the overlap in proteins between conditions



```{r, Frequency, echo=TRUE, warning=FALSE}


# Plot frequency of significant conditions per protein and the overlap in proteins between conditions

plot_cond(dep)

```


## results


```{r, Results, echo=TRUE, warning=FALSE}


# Generate a results table
data_results = get_results(dep)

# Number of significant proteins
cat('Number of significant abundance')
data_results %>% dplyr::filter(significant) %>% nrow()

# Column names of the results table
# colnames(data_results)


# https://stackoverflow.com/questions/12884695/collapse-rows-in-a-data-frame-using-r
# dt = pathwaysDF[, list(variables = paste('bin', collapse = " | ")), by = 'pID']
dt = plyr::ddply(pathwaysDF, 'pID', plyr::summarize, variables = paste(sort(bin), collapse = " | "))
# data.table::setDF(dt)
colnames(dt)[2] = 'BIN'
colnames(dt)[1] = 'pID'

colnames(data_results)[2] = 'pID'

# table(dt$pID %in% data_results$pID)
# table(data_results$pID %in% dt$pID)
# table(Prot$pID %in% data_results$pID)
# table(data_results$pID %in% Prot$pID)



data_results = data_results[, grep('pID|ratio|p\\.adj', colnames(data_results))]
rs = rowSums(data_results[, grep('p\\.adj', colnames(data_results))] < threshold, na.rm = TRUE)
data_results$diffAbundant = FALSE
data_results$diffAbundant[rs >= 1] = TRUE
# table(data_results$diffAbundant)

# Add annotations i any
data_results = merge(dt, data_results, by = 'pID', all.x = FALSE, all.y = TRUE)
# data_results = merge(data_results, Prot, by = 'pID', all.x = TRUE, all.y = TRUE)
# ....
  


# Write results
fpo = file.path('..', 'output')
fn = 'Proteomics-DA-DEP.txt'
data_results = data_results %>%
  dplyr::mutate(dplyr::across(where(is.numeric), ~ round(., digits = 2)))

write.table(data_results,
            file = file.path(fpo, fn),
            append = FALSE,
            quote = FALSE,
            sep = "\t",
            eol = "\n",
            na = "-",
            dec = ".",
            row.names = FALSE,
            col.names = TRUE,
            qmethod = c("escape"),
            fileEncoding = "UTF-8")




```


# session Info

```{r, sessionInfo}

sessionInfo()

```
