==> _p_ADAPTOmics/_I_Desiree/_S_multiOmics/_A_multiOmicsStat-R/scripts/02.2_merge-GSEA.Rmd <==
---
title: "02.2_merge-GSEA"
author: "zagor"
date: "`r Sys.Date()`"
output:
  html_document:
    fig_caption: yes
    self_contained: yes
    fig_width: 8
    fig_height: 9
    toc: yes
    toc_float: true
    code_folding: show
    toc_depth: 5
    number_sections: yes
    theme: flatly
    highlight: tango
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE, echo=FALSE, warning=FALSE}
knitr::opts_chunk$set(#dev = c('pdf', 'png'),  # this embeds pdf and crates scrolable blocks
                      dev = c('png'), 
                      fig.align = 'center', 
                      fig.height = 9, 
                      fig.width = 8 ,
                      warning = FALSE, message = FALSE
                      )
# options(knitr.table.format = "html")

```



```{r, echo=TRUE, warning=FALSE, message=FALSE}

rm(list = ls(all = TRUE))
gc()


set.seed(123456)

```

# gmt file

```{r}

fp = file.path('..', 'input')
fn = 'stu_StPanTr_3_CDS_2021-02-26_mapping_manual.gmt'
tmp = (readLines(file.path(fp, fn)))
n = stringr::str_count(tmp, pattern = '\t')
l = tmp[which(n == max(n))]
m = max(n+1)
temp = stringr::str_split_fixed(tmp, '\t', m)
gmm = as.data.frame(temp)
gmm = data.frame(gmm[, 1])
colnames(gmm)[1] = 'NAME'

rm(temp)
gc()

```

# GSEA java results

```{r, echo=TRUE, warning=FALSE, message=FALSE}

fp = file.path('..', 'other')
fl = list.files(fp, recursive = TRUE, pattern = 'gsea_report_for_.*\\.tsv')

cnt = 1
dff = NULL
gmmPlus = gmm

for (i in fl){
  
  # print(i)
  
  df = data.table::fread(file.path(fp, i), header = TRUE)
  data.table::setDF(df)
  ind = grep('MSigDB|GS DETAILS|V12', colnames(df))
  df = df[, -ind]
  n = stringr::str_count(df$`LEADING EDGE`, pattern = ',')
  m = max(n+1)
  df$`LEADING EDGE` = gsub('tags=|list=|signal=', '', df$`LEADING EDGE`)
  temp = stringr::str_split_fixed(df$`LEADING EDGE`, ',', m)
  colnames(temp) = c('tags', 'list', 'signal')
  ind = grep('LEADING EDGE', colnames(df))
  df = df[, -ind]
  df = cbind(df, temp)
  name = sub(pattern = "(.*)\\..*$", replacement = "\\1", basename(i))
  name = gsub('gsea_report_for_', '', name)
  print(name)
  name = gsub('S', '', name)
  name = paste(stringr::str_split_fixed(name, '_', 3)[1:2], collapse = '')
  
  if (!(cnt %% 2)) {
    dff = rbind(dff, df)
    colnames(dff)[2:ncol(dff)] = paste(colnames(dff), name, sep = '|')[2:ncol(dff)]
    # table(duplicated(dff$NAME))
    gmmPlus = merge(gmmPlus, dff, by = 'NAME', all.x = TRUE, all.y = TRUE)
    dff = NULL
  } else {
    dff = rbind(dff, df)
  }
  
  cnt = cnt + 1
  
  
}


rm(gmm)
gc()

```

# write table(s)

```{r, echo=TRUE, warning=FALSE, message=FALSE}

# remove empty
rs = rowSums(is.na(gmmPlus[, 2:ncol(gmmPlus)]))
# table(rs)
# table(rs == (ncol(gmmPlus) - 1))
ind = which(rs == (ncol(gmmPlus) - 1))
gmmPlus = gmmPlus[-ind, ]


# fpo = file.path('..', 'reports')
# fn = 'GSEA_all-cols.txt'
# write.table(gmmPlus,
#             file = file.path(fpo, fn),
#             append = FALSE,
#             quote = FALSE,
#             sep = "\t",
#             eol = "\n",
#             na = "-",
#             dec = ".",
#             row.names = FALSE,
#             col.names = TRUE,
#             qmethod = c("escape"),
#             fileEncoding = "UTF-8")



ind = c(grep('^NAME$', colnames(gmmPlus)),
        min(grep('SIZE', colnames(gmmPlus))),
        grep('^NES|^FDR|^tags', colnames(gmmPlus)))
gmmPlus = gmmPlus[, ind]
colnames(gmmPlus)[2] = 'SIZE'

# leading 0 for sort
colnames(gmmPlus) = gsub('1$', '01', colnames(gmmPlus))
colnames(gmmPlus) = gsub('7$', '07', colnames(gmmPlus))
colnames(gmmPlus) = gsub('8$', '08', colnames(gmmPlus))

ind = match(sort(colnames(gmmPlus)), colnames(gmmPlus))
gmmPlus = gmmPlus[, ind]
ind = grep('NAME|SIZE', colnames(gmmPlus))
gmmPlus = gmmPlus[, c(ind, setdiff(seq(1, ncol(gmmPlus), 1), ind))]


fpo = file.path('..', 'output')
fn = 'Proteomics-GSEA_few-cols.txt'
write.table(gmmPlus,
            file = file.path(fpo, fn),
            append = FALSE,
            quote = FALSE,
            sep = "\t",
            eol = "\n",
            na = "-",
            dec = ".",
            row.names = FALSE,
            col.names = TRUE,
            qmethod = c("escape"),
            fileEncoding = "UTF-8")



```

# session Info


```{r, echo=TRUE, warning=FALSE, message=FALSE}

sessionInfo()


```
