==> _p_ADAPTOmics/_I_Desiree/_S_multiOmics/_A_multiOmicsStat-R/scripts/02.3_plot-GSEA.Rmd <==
---
title: "01_gseaFromStats"
author: "zagor"
date: "`r Sys.Date()`"
output:
  html_document:
    fig_caption: yes
    self_contained: yes
    fig_width: 8
    fig_height: 6
    toc: yes
    toc_float: true
    code_folding: show
    toc_depth: 5
    number_sections: yes
    theme: flatly
    highlight: tango
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE, echo=FALSE, warning=FALSE}
knitr::opts_chunk$set(#dev = c('pdf', 'png'),  # this embeds pdf and crates scrolable blocks
                      dev = c('png'), 
                      fig.align = 'center', 
                      fig.height = 6, 
                      fig.width = 8 ,
                      warning = FALSE, message = FALSE
                      )
# options(knitr.table.format = "html")

```



```{r, echo=TRUE, warning=FALSE}

rm(list = ls(all = TRUE))
gc()


set.seed(123456)

`%!in%` = Negate(`%in%`)


```




# libraries

```{r, warning=FALSE, message=FALSE}

# devtools::install_github(repo = "https://github.com/martingarridorc/biokit")
library(biokit)

library(magrittr)
library(ggplot2)

library(RColorBrewer)

```



# bin selection and naming

```{r, echo = TRUE, warning=FALSE, message=FALSE}

fp = file.path('..', 'other')
fn = 'bins-to-plot.txt'

bins = data.table::fread(file.path(fp, fn), sep = '\t', header = FALSE)




```

# threshold

```{r, echo = TRUE, warning=FALSE, message=FALSE}

CutOff = 0.1
# CutOff = 0.05

```


```{r, echo = FALSE, warning=FALSE, message=FALSE}





my.gseaRes = function (gseaResDf, pCutoff, mytitle) {
  
  # i = as.numeric(gsub(' .*', '', gsub('\\..*', '', gseaResDf$pathway)))
  gseaResDf$order = unique(gseaResDf$pathway) # i
  gseaResDf = gseaResDf[order(gseaResDf$order, gseaResDf$pathway), ]
  gseaResDf = gseaResDf[, -ncol(gseaResDf)]
  gseaResDf$pathway = factor(gseaResDf$pathway, levels = rev(unique(gseaResDf$pathway)))
  
  r = range(gseaResDf$NES, na.rm = TRUE)
  m = max(abs(r))
  

    outP <- gseaResDf %>% 
      dplyr::mutate(status = ifelse(NES > 0, "Up", "Down")) #%>% 
      #subset(padj < pCutoff) 

      
      p = ggplot(outP,
             aes(x = comparison, 
                 y = pathway, 
                 fill = NES,
                 size = ifelse(-log10(padj) != 0, -log10(padj), 0), 
                 shape = ifelse(-log10(outP$padj) != 0, status, 1)
                 )) + 
      geom_point(alpha = ifelse(-log10(outP$padj) != 0, 0.75, 100)) + 
      scale_shape_manual(values = c(Up = 24, #21,
                                    Down = 25# 21#25
                                    )
                         ) +
      scale_fill_gradient2(high = "darkred",#("darkred"),
                           mid = "white",
                           low = "darkblue",#("darkblue"),
                           midpoint = 0,
                           limits = c(-m,m),
                           # n.breaks = 10,
                           breaks = c(-1, 0, 1),
                           na.value = 'white') +
      guides(x = guide_axis(angle = 90#,
                            #n.dodge = 
                              )) + 
      # theme_bw() + 
      theme_classic() + 
        ggtitle(paste0(mytitle#, " (FDR < ",
                       #pCutoff, ")"
                       )) +
      theme(
            axis.text.x = element_text(size = 12),
            axis.text.y = element_text(size = 12),
            axis.title=element_text(size=14,face="bold"),
        plot.title = element_text(hjust = 0.5, size=22)) +
        xlab('') + 
        ylab('') +
        # ylab('MapMan BIN') +
      theme(legend.position="none") +
        # adjust as needed
        # geom_vline(xintercept = c(2.5, 4.5, 6.5, 8.5), col = 'gray75') +
        # geom_vline(xintercept = c(4.5), col = 'gray40')
        # geom_vline(xintercept = c(4.5, 6.5, 8.5), col = 'gray75') 
        geom_vline(xintercept = c(2.5, 6.5, 8.5, 11.5), col = 'gray75') 
        # geom_vline(xintercept = c(4.5), col = 'gray75') # for HD
      
    return(p)

}


```


# NES and q-value

```{r, echo = TRUE, warning=FALSE, message=FALSE}

fp = file.path('..', 'output')
fn = 'Proteomics-GSEA_few-cols.txt'
gseaR = data.table::fread(file.path(fp, fn),
                          na.strings = '-')
gseaR = data.frame(lapply(gseaR, function(x) {
                  gsub('---', NA, x)
              }))


colnames(gseaR)[1] = c('pathway')
gseaR = gseaR[, c(1, grep('^NES|^FDR', colnames(gseaR)))]
gseaR = gseaR[!is.na(gseaR$pathway), ]


# colnames(gseaR) = gsub('\\.1', '_padj', colnames(gseaR))
i = grep('^NES', colnames(gseaR))
NES = tidyr::gather(gseaR[, c(1, i)], 
                    comparison, NES, colnames(gseaR)[min(i)]:colnames(gseaR)[max(i)], 
                    factor_key=TRUE)
i = grep('^FDR', colnames(gseaR))
padj = tidyr::gather(gseaR[, c(1, i)], 
                     comparison, padj, colnames(gseaR)[min(i)]:colnames(gseaR)[max(i)], 
                     factor_key=TRUE)

NES$comparison = gsub('NES\\.', '', NES$comparison)
padj$comparison = gsub('FDR.*\\.val\\.', '', padj$comparison)
NES$comparison = gsub('\\.', '-', NES$comparison)
padj$comparison = gsub('\\.', '-', padj$comparison)

gseaRes = merge(NES, padj, by = c('pathway', 'comparison'), all.x = TRUE, all.y = TRUE)
typeof(gseaRes$NES)
typeof(gseaRes$padj)
gseaRes$NES = as.numeric(gseaRes$NES)
gseaRes$padj = as.numeric(gseaRes$padj)

gseaRes = gseaRes[!duplicated(gseaRes), ]
gseaRes$padj[is.na(gseaRes$padj)] = 1


gseaRes$comparison = factor(gseaRes$comparison, levels = sort(unique(gseaRes$comparison)))

hist(gseaRes$NES, breaks = 100, main = 'NES')

```


# subset

```{r, echo = TRUE, warning=FALSE, message=FALSE}



# table(bins$V1 %in% mySubset$pathway)
mySubset = gseaRes[gseaRes$pathway %in% bins$V1, ]
# table(mySubset$padj < CutOff)

mySubset$NES[mySubset$padj >= CutOff] = NA
# range(mySubset$padj, na.rm = TRUE)
mySubset$padj[mySubset$padj == 0] = min(mySubset$padj[mySubset$padj != 0], na.rm = TRUE)/2

mySubset$comparison = droplevels(mySubset$comparison)
mySubset$NES[is.na(mySubset$NES)] = 0
mySubset$padj[mySubset$NES == 0] = 1

ind = match(mySubset$pathway, bins$V1)
mySubset$pathway = bins$V2[ind]
mySubset$pathway = factor(mySubset$pathway , levels = bins$V2)




```


# plot

```{r, echo = TRUE, warning=FALSE, message=FALSE}


p1 = my.gseaRes(gseaResDf = mySubset, 
                pCutoff = CutOff, 
                mytitle = '')



p1

```



# sessionInfo

```{r, echo = TRUE, warning=FALSE, message=FALSE}

sessionInfo()

```
