==> _p_ADAPTOmics/_I_Desiree/_S_multiOmics/_A_multiOmicsStat-R/scripts/03.2_Phenomics-invasive.Rmd <==
---
title: "03.2_Phenomics-invasive"
author: "zagor"
date: "`r Sys.Date()`"
output:
  html_document:
    fig_caption: yes
    self_contained: yes
    fig_width: 6
    fig_height: 5
    toc: true
    toc_float:
      toc_collapsed: false
    toc_depth: 5
    number_sections: yes
    theme: flatly
    highlight: tango
  pdf_document:
    toc: yes
    toc_depth: '5'
  word_document:
    toc: yes
    toc_depth: '5'
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(#dev = c('pdf', 'png'),  # this embeds pdf and crates scrolable blocks
                      dev = c('png'), 
                      fig.align = 'center', 
                      fig.height = 5, 
                      fig.width = 6 ,
                      warning = FALSE, message = FALSE
                      )

```



```{r, rm, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

rm(list = ls(all = TRUE))

gc()
gc()

seed = 123456
set.seed(seed)

```


# libraries


```{r, echo=TRUE, warning=FALSE, message=FALSE}


library(ggpubr)
library(ggplot2)

```

# palette

```{r}

mypallette = c('#7F7F7F', '#FFC000', '#C00000', '#ED7D31', '#4472C4') # , '#70AD47')
# pie(rep(1, length(mypallette)), col = mypallette, labels = c('C', 'D', 'H', 'HD', 'W' ))# , 'HDW'))

jitter = position_jitter(width = 0.1, height = 0.0)
dodge = position_dodge(width=0.5)

```


# tuber data


```{r}

# relative path to data
fp = file.path('..', 'input')
# file name
fn = 'data_targeted.xlsx'

data = openxlsx::read.xlsx(file.path(fp, fn), sheet = 'phenomics-invasive')
data.table::setDF(data)

data$Treatment = gsub('_.*', '', data$SampleID)
data = data[data$Treatment != 'HDW', ]



data$Time = as.numeric(gsub('S|P', '', stringr::str_split_i(data$SampleID, "_", i = 2)))

# unique(data[data$Time < 8, ]$Treatment)
# unique(data[data$Time > 14, ]$Treatment)

data$Treatment = as.factor(data$Treatment)



```


# No. of tubers and Total tubers weight

## plot and compare


```{r}

set.seed(seed)

my_comparisons <- list( c("C", "D"), 
                        c("C", "H"), 
                        c("C", "HD"), 
                        c("C", "W"))




ggviolin(data, x = "Treatment", y = "Num_Tubers", fill = "Treatment",
         palette = mypallette,
         add = "mean_sd",# "boxplot", add.params = list(fill = "white"), # 
         ylab = 'Number of tubers',
        xlab = 'Treatment'
        ) +
  stat_compare_means(comparisons = my_comparisons, method = "wilcox.test", 
                     method.args = list(alternative  = 'two.sided', paired = FALSE, exact = FALSE, correct = TRUE)
                     #,
                     #label = "p.signif", hide.ns = TRUE
                     ) + # Add significance levels
  stat_compare_means(method = 'kruskal.test', label.y = 0) +
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black")) +
  scale_y_continuous(breaks = seq(0, 20, by = 5)) +
  scale_x_discrete(label = c('C', 'D', 'H', 'HD', 'W'))






ggviolin(data, x = "Treatment", y = "Total_Tubers_Weight", fill = "Treatment",
         palette = mypallette,
         add = # "boxplot", add.params = list(fill = "white"),  # 
         "mean_sd",
         ylab = 'Total Tubers Weight',
        xlab = 'Treatment'
        ) +
  stat_compare_means(comparisons = my_comparisons, 
                     method = "wilcox.test", 
                     method.args = list(alternative  = 'two.sided', paired = FALSE, exact = FALSE, correct = TRUE)
                     #,
                     #label = "p.signif", hide.ns = TRUE
                     ) + # Add significance levels
  stat_compare_means(method = 'kruskal.test', label.y = 0) +
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black")) +
  scale_y_continuous(breaks = seq(0, 150, by = 25)) +
  scale_x_discrete(label = c('C', 'D', 'H', 'HD', 'W'))


```

# Fresh and Dry weight

## plot

```{r}

ggdotplot(data[data$Time <= 14, ], x = "Time", y = "FW_SH_Total", fill = "Treatment",
         palette = mypallette,
         add = "mean_sd",# "boxplot", add.params = list(fill = "white"), # 
         ylab = 'Fresh weight',
        xlab = 'Days',
        size = 0.75,
        mean.point.size = 0.75
        ) +
  facet_wrap(~Treatment, ncol = 5) + 
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black"))


ggdotplot(data[data$Time <= 14, ], x = "Time", y = "DW_SH_Total", fill = "Treatment",
         palette = mypallette,
         add = "mean_sd",# "boxplot", add.params = list(fill = "white"), # 
         ylab = 'Dry weight',
        xlab = 'Days',
        size = 0.75,
        mean.point.size = 0.75
        ) +
  facet_wrap(~Treatment, ncol = 5) + 
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black"))


```

## compare

```{r}

set.seed(seed)

g1 = ggboxplot(data[data$Time <= 7, ], x = "Treatment", y = "FW_SH_Total", fill = "Treatment",
         palette = mypallette[c(1,3,5)],
         add = "mean_sd",# "boxplot", add.params = list(fill = "white"), # 
         ylab = 'Fresh weight',
        xlab = 'Treatment'
        ) +
  stat_compare_means(comparisons = list(my_comparisons[[2]], my_comparisons[[4]]), method = "wilcox.test",
                     method.args = list(alternative  = 'two.sided', paired = FALSE, exact = FALSE, correct = TRUE)
                     , label = "p.signif", hide.ns = FALSE
                     ,step.increase = 0.15
                     ) + # Add significance levels
  stat_compare_means(method = 'kruskal.test', label.y = 0) +
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black"),
        strip.background = element_blank(),
        strip.text = element_text(size = 14)) +
  scale_x_discrete(label = c('C', 'H', 'W')) +
  facet_wrap(~Time, drop = TRUE)



g2 = ggboxplot(data[data$Time <= 14 & data$Time >=8, ], x = "Treatment", y = "FW_SH_Total", fill = "Treatment",
         palette = mypallette,
         add = "mean_sd",# "boxplot", add.params = list(fill = "white"), # 
         ylab = 'Fresh weight',
        xlab = 'Treatment'
        ) +
  stat_compare_means(comparisons = my_comparisons, method = "wilcox.test",
                     method.args = list(alternative  = 'two.sided', paired = FALSE, exact = FALSE, correct = TRUE)
                     , label = "p.signif", hide.ns = FALSE
                     ,step.increase = 0.15
                     ) + # Add significance levels
  stat_compare_means(method = 'kruskal.test', label.y = 0) +
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black"),
        strip.background = element_blank(),
        strip.text = element_text(size = 14)) +
  scale_x_discrete(label = c('C', 'D', 'H', 'HD', 'W')) +
  facet_wrap(~Time, drop = TRUE)


# ggarrange(g1, g2,  
#           labels = c("", ""),
#           ncol = 1, nrow = 2)



g3 = ggboxplot(data[data$Time <= 7, ], x = "Treatment", y = "DW_SH_Total", fill = "Treatment",
         palette = mypallette[c(1,3,5)],
         add = "mean_sd",# "boxplot", add.params = list(fill = "white"), # 
         ylab = 'Dry weight',
        xlab = 'Treatment'
        ) +
  stat_compare_means(comparisons = list(my_comparisons[[2]], my_comparisons[[4]]), method = "wilcox.test",
                     method.args = list(alternative  = 'two.sided', paired = FALSE, exact = FALSE, correct = TRUE)
                     , label = "p.signif", hide.ns = FALSE
                     ,step.increase = 0.15
                     ) + # Add significance levels
  stat_compare_means(method = 'kruskal.test', label.y = 0) +
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black"),
        strip.background = element_blank(),
        strip.text = element_text(size = 14)) +
  scale_x_discrete(label = c('C', 'H', 'W')) +
  facet_wrap(~Time, drop = TRUE)



g4 = ggboxplot(data[data$Time <= 14 & data$Time >=8, ], x = "Treatment", y = "DW_SH_Total", fill = "Treatment",
         palette = mypallette,
         add = "mean_sd",# "boxplot", add.params = list(fill = "white"), # 
         ylab = 'Dry weight',
        xlab = 'Treatment'
        ) +
  stat_compare_means(comparisons = my_comparisons, method = "wilcox.test",
                     method.args = list(alternative  = 'two.sided', paired = FALSE, exact = FALSE, correct = TRUE)
                     , label = "p.signif", hide.ns = FALSE
                     ,step.increase = 0.15
                     ) + # Add significance levels
  stat_compare_means(method = 'kruskal.test', label.y = 0) +
  theme(legend.position="none") +
  theme(axis.title.x = element_text(size=12, face="bold", color = "black"),
        axis.title.y = element_text(size=11, face="bold", color = "black"),
        strip.background = element_blank(),
        strip.text = element_text(size = 14)) +
  scale_x_discrete(label = c('C', 'D', 'H', 'HD', 'W')) +
  facet_wrap(~Time, drop = TRUE)

# ggarrange(g3, g4,  
#           labels = c("", ""),
#           ncol = 1, nrow = 2)

plot(g1)
plot(g3)
plot(g2)
plot(g4)


```


# session Info

```{r, echo=TRUE, warning=FALSE, message=FALSE}


sessionInfo()

```
