==> _p_ADAPTOmics/_I_Desiree/_S_multiOmics/_A_multiOmicsStat-R/scripts/04.1_targeted-logFC.Rmd <==
---
title: "04.1_targeted-logFC"
author: "zagor"
date: "`r Sys.Date()`"
output:
  html_document:
    fig_caption: yes
    self_contained: yes
    fig_width: 7
    fig_height: 6
    toc: true
    toc_float:
      toc_collapsed: false
    toc_depth: 5
    number_sections: yes
    theme: flatly
    highlight: tango
  pdf_document:
    toc: yes
    toc_depth: '5'
  word_document:
    toc: yes
    toc_depth: '5'
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(#dev = c('pdf', 'png'),  # this embeds pdf and crates scrolable blocks
                      dev = c('png'), 
                      fig.align = 'center', 
                      fig.height = 6, 
                      fig.width = 7 ,
                      warning = FALSE, message = FALSE
                      )

```



```{r, rm, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

rm(list = ls(all = TRUE))

gc()
gc()

seed = 123456
set.seed(seed)

```


# libraries

```{r, libs, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

library(scales)
library(ggplot2)
# packageVersion("ggplot2")
library(magrittr)
# library(crayon)
library(rstatix)
# library(ggpubr)
# library(Compositional)
# library(equalCovs)
# library(psych)
library(corrplot) 
# library(pdist)
library(RColorBrewer)
library(plotly)
library(heatmaply)
# library(Hmisc)
library(htmlwidgets)
library(viridis)
library(ggh4x)


`%nin%` = Negate(`%in%`)


```

# f-ctions


```{r, t-p, echo=FALSE, results='hide', warning=FALSE, message=FALSE}


####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####

my.customised.t.test <- function(Ctrl, Stress, StressC, CtrlC, genotype){
  

  set.seed(seed)
  
  Ctrl$Stress = unique(Ctrl$mygroup)
  stress$Stress = unique(stress$mygroup)
  
  
  mydata.long = rbind(Ctrl, stress)
  
  mydata.long = mydata.long[mydata.long$Genotype == genotype, ]
  mydata.long$measurement[mydata.long$measurement == 0] = NA # LOD
  mydata.long$log2 = log2(mydata.long$measurement)
  mydata.long$molecule = as.factor(mydata.long$molecule)


    stat.test <- mydata.long %>%
    dplyr::group_by(Genotype, molecule) %>%
    rstatix::t_test(log2 ~ Stress)
    

    if (any(stat.test$group1 %nin% Ctrl$Stress & stat.test$group2 %nin%  stress$Stress)) {
      tmp = stat.test$group1
      stat.test$group1 = stat.test$group2
      stat.test$group2 = tmp
    }

  
  return(stat.test)
  
}


####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####

my.logFC <- function(Ctrl, stress, StressC, CtrlC, genotype){
    

    Ctrl$Stress = StressC
    stress$Stress = CtrlC
    

    Ctrl = Ctrl[Ctrl$Genotype == genotype, ]
    stress = stress[stress$Genotype == genotype, ]
    
    stress$measurement[stress$measurement == 0] = NA # LOD
    Ctrl$measurement[Ctrl$measurement == 0] = NA # LOD

    

    Ctrl$log2 = log2(Ctrl$measurement)
    stress$log2 = log2(stress$measurement)
  
    
      Ctrl.mean =  Ctrl %>%
        dplyr::group_by(molecule) %>%
        dplyr::summarise_at(vars(log2), list(Ctrl.mean = ~ mean(., na.rm = TRUE)))
      
      stress.mean =  stress %>%
        dplyr::group_by(molecule) %>%
        dplyr::summarise_at(vars(log2), list(stress.mean = ~ mean(., na.rm = TRUE)))
      
            
      
      log2FC = stress.mean$stress.mean - Ctrl.mean$Ctrl.mean
      
      stress.mean = cbind(stress.mean, log2FC)
      colnames(stress.mean)[ncol(stress.mean)] = 'log2FC'

      tmp = merge(Ctrl.mean, stress.mean, by = c('molecule'))
      tmp$Genotype = genotype
      tmp$stressGroup = unique(stress$mygroup)
      tmp$controlGroup = unique(Ctrl$mygroup)


    
    return(tmp)

    
    

}


####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####  ####

```




```{r, summarySE, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

# adapted from <https://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/>

## Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
##   data: a data frame.
##   measurevar: the name of a column that contains the variable to be summariezed
##   groupvars: a vector containing names of columns that contain grouping variables
##   na.rm: a boolean that indicates whether to ignore NA's
##   conf.interval: the percent range of the confidence interval (default is 95%)
summarySE = function(data=NULL, 
                     measurevar, 
                     groupvars=NULL, 
                     na.rm=TRUE,
                     conf.interval=.95, 
                     .drop=TRUE) {


    # New version of length which can handle NA's: if na.rm==T, don't count them
    length2 <- function (x, na.rm=FALSE) {
        if (na.rm) sum(!is.na(x))
        else       length(x)
    }

    # This does the summary. For each group's data frame, return a vector with
    # N, mean, and sd
    datac <- plyr::ddply(data, groupvars, .drop=.drop,
      .fun = function(xx, col) {
        c(N    = length2(xx[[col]], na.rm=na.rm),
          mean = mean   (xx[[col]], na.rm=na.rm),
          sd   = sd     (xx[[col]], na.rm=na.rm)
        )
      },
      measurevar
    )

    # Rename the "mean" column    
    # datac <- rename(datac, c("mean" = measurevar))

    datac$se <- datac$sd / sqrt(datac$N)  # Calculate standard error of the mean

    # Confidence interval multiplier for standard error
    # Calculate t-statistic for confidence interval: 
    # e.g., if conf.interval is .95, use .975 (above/below), and use df=N-1
    ciMult <- qt(conf.interval/2 + .5, datac$N-1)
    datac$ci <- datac$se * ciMult
    
    datac$cv = datac$sd/datac$mean
    

    return(datac)
    
}

```


```{r, panel.cor, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

## c/p from stack
# Function to add correlation coefficients
# panel.cor <- function(x, y, digits = 2, prefix = "", cex.cor, ...) {
#   usr <- par("usr")
#   #on.exit(par(usr))
#   par(usr = c(0, 1, 0, 1))
#   Cor <- (cor(x, y, use = 'na.or.complete')) # Remove abs function if desired
#   txt <- paste0(prefix, format(c(Cor, 0.123456789), digits = digits)[1])
#   if(missing(cex.cor)) {
#     cex.cor <- 0.4 / strwidth(txt)
#   }
#   text(0.5, 0.5, txt,
#        # cex = 1 + cex.cor * 0.75, #Cor) # Resize the text by level of correlation
#        cex = 1,
#   col = ifelse(as.numeric(txt) < 0, 'darkblue', 'darkred'))
# }



pairs.panels = function (x, smooth = TRUE, scale = FALSE, density = TRUE, ellipses = TRUE, 
                         digits = 2, method = "pearson", pch = 20, lm = FALSE, cor = TRUE, 
                         jiggle = FALSE, factor = 2, hist.col = "cyan", show.points = TRUE, 
                         rug = TRUE, breaks = "Sturges", cex.cor = 1, wt = NULL, smoother = FALSE, 
                         stars = FALSE, ci = FALSE, alpha = 0.05, hist.border = "black", 
                         cccex = 1,
                         ...) {
  
    "panel.hist.density" <- function(x, ...) {
        usr <- par("usr")
        on.exit(par("usr"))
        par(usr = c(usr[1], usr[2], 0, 1.5))
        tax <- table(x)
        if (length(tax) < 11) {
            breaks <- as.numeric(names(tax))
            y <- tax/max(tax)
            interbreak <- min(diff(breaks)) * (length(tax) - 
                1)/41
            rect(breaks - interbreak, 0, breaks + interbreak, 
                y, col = hist.col, border = hist.border)
        }
        else {
            h <- hist(x, breaks = breaks, plot = FALSE)
            breaks <- h$breaks
            nB <- length(breaks)
            y <- h$counts
            y <- y/max(y)
            rect(breaks[-nB], 0, breaks[-1], y, col = hist.col, 
                border = hist.border)
        }
        if (density) {
            tryd <- try(d <- density(x, na.rm = TRUE, bw = "nrd", 
                adjust = 1.2), silent = TRUE)
            if (!inherits(tryd, "try-error")) {
                d$y <- d$y/max(d$y)
                lines(d)
            }
        }
        if (rug) 
            rug(x)
    }
    "panel.cor" <- function(x, y, prefix = "", ...) {
        usr <- par("usr")
        on.exit(par("usr"))
        par(usr = c(0, 1, 0, 1))
        if (is.null(wt)) {
            r <- cor(x, y, use = "pairwise", method = method)
        }
        else {
            r <- cor.wt(data.frame(x, y), w = wt[, c(1:2)])$r[1,
                2]
        }
        txt <- format(c(round(r, digits), 0.123456789), digits = digits)[1]
        txt <- paste(prefix, txt, sep = "")
        if (stars) {
            pval <- psych::r.test(sum(!is.na(x * y)), r)$p
            symp <- symnum(pval, corr = FALSE, cutpoints = c(0,
                0.001, 0.01, 0.05, 1), symbols = c("***", "**",
                "*", " "), legend = FALSE)
            txt <- paste0(txt, symp)
        }
        cex <- cex.cor # * 0.8/(max(strwidth("0.12***"), strwidth(txt)))
        if (scale) {
            cex1 <- cex * abs(r)
            if (cex1 < 0.25)
                cex1 <- 0.25
            text(0.5, 0.5, txt,
                 cex = cccex,
                 col = ifelse(!grepl('\\*', txt), 
                              'grey', 
                              ifelse(as.numeric(gsub('\\*', '', txt)) < 0, 
                                     'darkblue', 
                                     'darkred'))
                 )
        }
        else {
            text(0.5, 0.5, txt,
                 cex = cccex,
                 col = ifelse(!grepl('\\*', txt), 
                              'grey', 
                              ifelse(as.numeric(gsub('\\*', '', txt)) < 0, 
                                     'darkblue', 
                                     'darkred'))
                 )
            
        }
    }
    "panel.smoother" <- function(x, y, pch = par("pch"), col.smooth = "red", 
        span = 2/3, iter = 3, ...) {
        xm <- mean(x, na.rm = TRUE)
        ym <- mean(y, na.rm = TRUE)
        xs <- sd(x, na.rm = TRUE)
        ys <- sd(y, na.rm = TRUE)
        r = cor(x, y, use = "pairwise", method = method)
        if (jiggle) {
            x <- jitter(x, factor = factor)
            y <- jitter(y, factor = factor)
        }
        if (smoother) {
            smoothScatter(x, y, add = TRUE, nrpoints = 0)
        }
        else {
            if (show.points) 
                points(x, y, pch = pch, ...)
        }
        ok <- is.finite(x) & is.finite(y)
        if (any(ok)) {
            if (smooth & ci) {
                lml <- loess(y ~ x, degree = 1, family = "symmetric")
                tempx <- data.frame(x = seq(min(x, na.rm = TRUE), 
                  max(x, na.rm = TRUE), length.out = 47))
                pred <- predict(lml, newdata = tempx, se = TRUE)
                if (ci) {
                  upperci <- pred$fit + confid * pred$se.fit
                  lowerci <- pred$fit - confid * pred$se.fit
                  polygon(c(tempx$x, rev(tempx$x)), c(lowerci, 
                    rev(upperci)), col = adjustcolor("light grey", 
                    alpha.f = 0.8), border = NA)
                }
                lines(tempx$x, pred$fit, col = col.smooth, ...)
            }
            else {
                if (smooth) 
                  lines(stats::lowess(x[ok], y[ok], f = span, 
                    iter = iter), col = col.smooth)
            }
        }
        if (ellipses) 
            draw.ellipse(xm, ym, xs, ys, r, col.smooth = col.smooth, 
                ...)
    }
    "panel.lm" <- function(x, y, pch = par("pch"), col.lm = "red", 
        ...) {
        ymin <- min(y)
        ymax <- max(y)
        xmin <- min(x)
        xmax <- max(x)
        ylim <- c(min(ymin, xmin), max(ymax, xmax))
        xlim <- ylim
        if (jiggle) {
            x <- jitter(x, factor = factor)
            y <- jitter(y, factor = factor)
        }
        if (smoother) {
            smoothScatter(x, y, add = TRUE, nrpoints = 0)
        }
        else {
            if (show.points) {
                points(x, y, pch = pch, ylim = ylim, xlim = xlim, 
                  ...)
            }
        }
        ok <- is.finite(x) & is.finite(y)
        if (any(ok)) {
            lml <- lm(y ~ x)
            if (ci) {
                tempx <- data.frame(x = seq(min(x, na.rm = TRUE), 
                  max(x, na.rm = TRUE), length.out = 47))
                pred <- predict.lm(lml, newdata = tempx, se.fit = TRUE)
                upperci <- pred$fit + confid * pred$se.fit
                lowerci <- pred$fit - confid * pred$se.fit
                polygon(c(tempx$x, rev(tempx$x)), c(lowerci, 
                  rev(upperci)), col = adjustcolor("light grey", 
                  alpha.f = 0.8), border = NA)
            }
            if (ellipses) {
                xm <- mean(x, na.rm = TRUE)
                ym <- mean(y, na.rm = TRUE)
                xs <- sd(x, na.rm = TRUE)
                ys <- sd(y, na.rm = TRUE)
                r = cor(x, y, use = "pairwise", method = method)
                draw.ellipse(xm, ym, xs, ys, r, col.smooth = col.lm, 
                  ...)
            }
            abline(lml, col = col.lm, ...)
        }
    }
    "draw.ellipse" <- function(x = 0, y = 0, xs = 1, ys = 1, 
        r = 0, col.smooth, add = TRUE, segments = 51, ...) {
        angles <- (0:segments) * 2 * pi/segments
        unit.circle <- cbind(cos(angles), sin(angles))
        if (!is.na(r)) {
            if (abs(r) > 0) 
                theta <- sign(r)/sqrt(2)
            else theta = 1/sqrt(2)
            shape <- diag(c(sqrt(1 + r), sqrt(1 - r))) %*% matrix(c(theta, 
                theta, -theta, theta), ncol = 2, byrow = TRUE)
            ellipse <- unit.circle %*% shape
            ellipse[, 1] <- ellipse[, 1] * xs + x
            ellipse[, 2] <- ellipse[, 2] * ys + y
            if (show.points) 
                points(x, y, pch = 19, col = col.smooth, cex = 1.5)
            lines(ellipse, ...)
        }
    }
    "panel.ellipse" <- function(x, y, pch = par("pch"), col.smooth = "red", 
        ...) {
        segments = 51
        usr <- par("usr")
        on.exit(par("usr"))
        par(usr = c(usr[1] - abs(0.05 * usr[1]), usr[2] + abs(0.05 * 
            usr[2]), 0, 1.5))
        xm <- mean(x, na.rm = TRUE)
        ym <- mean(y, na.rm = TRUE)
        xs <- sd(x, na.rm = TRUE)
        ys <- sd(y, na.rm = TRUE)
        r = cor(x, y, use = "pairwise", method = method)
        if (jiggle) {
            x <- jitter(x, factor = factor)
            y <- jitter(y, factor = factor)
        }
        if (smoother) {
            smoothScatter(x, y, add = TRUE, nrpoints = 0)
        }
        else {
            if (show.points) {
                points(x, y, pch = pch, ...)
            }
        }
        angles <- (0:segments) * 2 * pi/segments
        unit.circle <- cbind(cos(angles), sin(angles))
        if (!is.na(r)) {
            if (abs(r) > 0) 
                theta <- sign(r)/sqrt(2)
            else theta = 1/sqrt(2)
            shape <- diag(c(sqrt(1 + r), sqrt(1 - r))) %*% matrix(c(theta, 
                theta, -theta, theta), ncol = 2, byrow = TRUE)
            ellipse <- unit.circle %*% shape
            ellipse[, 1] <- ellipse[, 1] * xs + xm
            ellipse[, 2] <- ellipse[, 2] * ys + ym
            points(xm, ym, pch = 19, col = col.smooth, cex = 1.5)
            if (ellipses) 
                lines(ellipse, ...)
        }
    }
    old.par <- par(no.readonly = TRUE)
    on.exit(par(old.par))
    if (missing(cex.cor)) 
        cex.cor <- 1
    for (i in 1:ncol(x)) {
        if (is.character(x[[i]])) {
            x[[i]] <- as.numeric(as.factor(x[[i]]))
            colnames(x)[i] <- paste(colnames(x)[i], "*", sep = "")
        }
    }
    n.obs <- nrow(x)
    confid <- qt(1 - alpha/2, n.obs - 2)
    if (!lm) {
        if (cor) {
            pairs(x, diag.panel = panel.hist.density, upper.panel = panel.cor, 
                lower.panel = panel.smoother, pch = pch, ...)
        }
        else {
            pairs(x, diag.panel = panel.hist.density, upper.panel = panel.smoother, 
                lower.panel = panel.smoother, pch = pch, ...)
        }
    }
    else {
        if (!cor) {
            pairs(x, diag.panel = panel.hist.density, upper.panel = panel.lm, 
                lower.panel = panel.lm, pch = pch, ...)
        }
        else {
            pairs(x, diag.panel = panel.hist.density, upper.panel = panel.cor, 
                lower.panel = panel.lm, pch = pch, ...)
        }
    }
}


```


```{r, my.pairs.panels, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

my.pairs.panels = function(df, mytitle, mycond, cccex) {
  
    pairs.panels(df,
                 smooth = FALSE,      # If TRUE, draws loess smooths
                 scale = FALSE, #TRUE,      # If TRUE, scales the elation text font
                 density = TRUE,     # If TRUE, adds density plots and histograms
                 ellipses = FALSE,    # If TRUE, draws ellipses
                 method = 'pearson', # elation method (also "spearman" or "kendall")
                 pch = 21,           # pch symbol
                 lm = TRUE,         # If TRUE, plots linear fit rather than the LOESS (smoothed) fit
                 cor = TRUE,         # If TRUE, reports elations
                 jiggle = FALSE,     # If TRUE, data points are jittered
                 factor = 2,         # Jittering factor
                 hist.col = 4,       # Histograms color
                 stars = TRUE,       # If TRUE, adds significance level with stars
                 ci = TRUE,          # If TRUE, adds confidence intervals
                 bg = extend.palette[extend.palette$Genotype == mytitle & extend.palette$group %in% mycond, ]$colour,
                 col = extend.palette[extend.palette$Genotype == mytitle & extend.palette$group %in% mycond, ]$colour,
                 cex.labels = cccex, # 1, # diagonal
                 cex = cccex, #4,
                 main = paste(mytitle, paste(mycond, collapse = ', ')))
  
  

}


```

```{r, cor.test.p, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

# Compute correlation p-values
cor.test.p = function(x){
  
  set.seed(seed)
  
    FUN = function(x, y) cor.test(x, y)[["p.value"]]
    z = outer(
      colnames(x), 
      colnames(x), 
      Vectorize(function(i,j) FUN(x[,i], x[,j]))
    )
    dimnames(z) = list(colnames(x), colnames(x))
    z
}


```


```{r, my.heatmaply, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

my.heatmaply = function(df, mytitle, mycond, subdir) {
  
  cor.coef = cor(df, use = 'pairwise.complete.obs')
  p = cor.test.p(df)
  p[p == 0] = min(p[p != 0])
  
  diag(p) = min(p[lower.tri(p, diag=FALSE)]) # tune down diagonal
  p[p >= 0.05] = 1
  point_size_mat = -log10(p)

  heatmaply_cor(
    cor.coef,
    node_type = "scatter",
    point_size_mat = point_size_mat, 
    point_size_name = "-log10(p-value)",
    #label_names = c("x", "y", "Correlation")
    main = paste(mytitle, paste(mycond, collapse = ', '), '(sig PCC)') ,
    colors = rev(my.heatmap.col1),
    file = file.path('..', 'reports', subdir, paste0('heatmaply_PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".html")),
    dendrogram = 'none'
  )
  
}

```


```{r, myggplot, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

myggplot = function(data.SE, title, cnt, lncol, lims, subdir) {

  myplotY = ggplot(data.SE, aes(x = PostInductionDay , 
                                y = mean, 
                                group = Condition,
                                shape = Condition,
                                fill = Condition,
                                colour = Condition)) + 
      facet_wrap(~ molecule, ncol = ncol,
             strip.position = "top",
             scales = 'free_y',
             drop = FALSE
             ) +
         # facet_grid(rows = vars(molecule), scales = "free_y",
         #           drop = FALSE) +
    ggh4x::facetted_pos_scales(y = lims) + 

      geom_point(data = data.SE,
                 aes(x = PostInductionDay , 
                     y = mean),
                 size = 2.5, 
                 colour = 'black'
                 ) + 
      geom_line(data = data.SE,
                aes(color = Condition, 
                    group = Condition,
                    linetype = Condition), 
                linewidth = 1) +
      geom_ribbon(data = data.SE, aes(ymin=mean-se, ymax=mean+se,
                                    x = PostInductionDay ,
                                    group = Condition,
                                    fill = Condition,
                                    linetype = Condition), 
                  alpha=0.1) +

      ggtitle(title)  +
    theme_classic() +
      scale_colour_manual(name = 'Legend',
                          values = mypallette) +
      scale_fill_manual(name = 'Legend',
                        values = mypallette
                        ) + 
    # guides(fill="none") +
        labs(x = '', 
             y = 'Relative molecule abundance [+/- SE]') + 
        scale_y_continuous(# trans = log2_trans(), 
                           labels = scales::comma#,
                           #labels =  trans_format("log2", math_format(2^.x))#,
                           # breaks = mybreaks,
                           # limits = mylimits
                           # labels = fancy_scientific
                           ) +
    # scale_x_discrete(labels = cd, breaks = cd) + 
    # https://datavizpyr.com/how-to-dodge-overlapping-text-on-x-axis-labels-in-ggplot2/
    # scale_x_continuous(labels = cd, breaks = cd,
    #                    guide = guide_axis(n.dodge=2)) + 
      theme(axis.text.x = element_text (angle = 0, # 45, 
                                        vjust = 1, 
                                        hjust = 1, 
                                        size = 12),
            axis.text.y = element_text (angle = 0, 
                                        vjust = 1, 
                                        hjust = 1, 
                                        size = 12),
            axis.title.x = element_text (angle = 0, # 45, 
                                        #vjust = 1, 
                                        #hjust = 1, 
                                        size = 12,
                                        face = "bold"),
            axis.title.y = element_text (angle = 90, 
                                        # vjust = 1, 
                                        # hjust = 1, 
                                        size = 12,
                                        face = "bold"),
            axis.text = element_text( angle = 90 ),
            panel.grid.major = element_blank(),
            panel.grid.minor = element_blank(),
            panel.border = element_blank(),
            panel.background = element_blank(),
            legend.position = "bottom",
            legend.background = element_rect(fill=NULL, 
                                             size=0.1, linetype="solid"),
            legend.title = element_text(color = NULL,
                                        size = 12,
                                        face = "bold"),
            legend.text = element_text( size = 12),
            strip.text = element_text(size = 12)) + 
  theme(# legend.position = 'bottom', # c(0.3, 1.0), 
        legend.box = "horizontal",
        # https://www.statology.org/ggplot2-legend-size/
        legend.key.width = unit(2, 'cm'),
        # https://stackoverflow.com/questions/54438495/shift-legend-into-empty-facets-of-a-faceted-plot-in-ggplot2
        legend.position = c(1, 0),
        legend.justification = c(1, 0)) +
  guides(colour = guide_legend(ncol = lncol)) +
    guides(linetype = guide_legend(ncol = lncol)) +

  scale_shape_manual(values=c(21,21,21,21,21,21)) +
  # guides(fill = "none") +
  guides(shape = "none") +
  theme(panel.spacing.y = unit(1.5, "lines")) +
  # scale_linetype_manual(values=c("solid", "twodash", "dotted",
  #                                "dot-dash", "longdash", "twodash"))
  geom_vline(xintercept=2, linetype='dashed', color='grey90', size=0.5) +
  geom_vline(xintercept=4, linetype='dashed', color='grey90', size=0.5) +
  # https://stackoverflow.com/questions/37140266/how-to-merge-color-line-style-and-shape-legends-in-ggplot
  labs(color  = "Legend", linetype = "Legend", shape = "Legend")
  
  
  
  fng = paste0(stringr::str_pad(cnt, 3, pad = "0"), '_', title, "_L.pdf")
  ggsave(file.path('..', 'reports', subdir, fng), 
         width = 16, height = 12, units = "in")
  fng = paste0(stringr::str_pad(cnt, 3, pad = "0"), '_', title, "_L.svg")
  ggsave(file.path('..', 'reports', subdir, fng),
         width = 16, height = 12, units = "in")
  
  return(myplotY)
  

}

```


```{r, my.pheatmap, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

my.pheatmap = function(df, mytitle, mycond, subdir){
 
  color.divisions = length(my.heatmap.col1)
  p = cor.test.p(df)
  p[p<0.05] = '*'
  p[p >= 0.05] = ''
  df.cor = cor(df, use = 'pairwise.complete.obs')
  p[df.cor < 0.75 & df.cor > -0.75] = ''

  
  pheatmap::pheatmap(df.cor, 
                     cluster_row = FALSE, cluster_cols = FALSE,
                     # legend_breaks = seq(-1, 1, 0.2),
                     breaks = seq(-1,1, length.out=(color.divisions + 1)),
                     color = rev(my.heatmap.col2),
                     main = paste0(mytitle, ' ', paste(mycond, collapse = '-')),
                     filename = file.path('..', 'reports', subdir, paste0('pheatmap_PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
                     cellwidth = 10, cellheight = 10,
                    display_numbers = p,
                    width = 8, height = 8)

}


```


```{r, my.pheatmap.logFC, echo=FALSE, results='hide', warning=FALSE, message=FALSE}

my.pheatmap.logFC = function(df, forHeatmap, cluster_rows = cluster_rows, cluster_cols = cluster_cols,
                             main, subdir, width = width, height = height, cnt = cnt){
  
  logFC = df[, grep('molecule$|stressGroup$|controlGroup$|log2FC', colnames(df))]
  logFC$comparison = paste(logFC$stressGroup, logFC$controlGroup, sep = '-')
  logFC = logFC[, -c(2,3)]
  
  p = df[, grep('molecule$|stressGroup$|controlGroup$|^p$', colnames(df))]
  p$comparison = paste(p$stressGroup, p$controlGroup, sep = '-')
  p = p[, -c(2,3)]
  
  logFC$comparison = factor(logFC$comparison, levels = forHeatmap$comparison )
  logFC = logFC[order(logFC$molecule, logFC$comparison), ]
  p$comparison = factor(p$comparison, levels = forHeatmap$comparison )
  p = p[order(p$molecule, p$comparison), ]
  
  logFC.w = tidyr::spread(logFC, comparison, log2FC)
  p.w = tidyr::spread(p, comparison, p)
  
  
  logFC.w[is.na(logFC.w)] = 0
  rownames(logFC.w) = logFC.w[, 1]
  logFC.w = logFC.w[, -1]
  
  p.w[is.na(p.w)] = 1
  rownames(p.w) = p.w[, 1]
  p.w = p.w[, -1]
  
  # all(rownames(p.w) == rownames(logFC.w))
  # all(colnames(p.w) == colnames(logFC.w))
  
  ind = match(sort(match(rownames(logFC.w), path$name)), match(rownames(logFC.w), path$name))
  logFC.w = logFC.w[ind, ]
  ind = match(sort(match(colnames(logFC.w), forHeatmap$comparison)), match(colnames(logFC.w), forHeatmap$comparison))
  logFC.w = logFC.w[, ind]
  
  ind = match(sort(match(rownames(p.w), path$name)), match(rownames(p.w), path$name))
  p.w = p.w[ind, ]
  ind = match(sort(match(colnames(p.w), forHeatmap$comparison)), match(colnames(p.w), forHeatmap$comparison))
  p.w = p.w[, ind]
  
  
  
  

  paletteLength = 100
  myColor = colorRampPalette(c("navy", "white", "darkred"))(paletteLength)
  
  
  
  rg = max(abs(logFC.w))
  myBreaks = seq(-rg, rg, length.out = 100)
  
  
  phfn = file.path('..', 'reports', subdir, paste0(stringr::str_pad(cnt, 3, pad = "0"), 
                                                   '_pheatmap_logFC_', main, ".pdf"))
  ph = pheatmap::pheatmap(logFC.w, cluster_rows = FALSE, cluster_cols = FALSE,
                     color=myColor, breaks=myBreaks,
                     main = main,
                     display_numbers = matrix(ifelse(p.w >= 0.05, "", '*'), nrow(p.w)),
                     filename = phfn)$gtable
  ggsave(filename = phfn,
         width = width, height = height, units = "in",
         plot=ph)
  
  dev.off()
  grid::grid.draw(ph)
  
  
  

  phfn = file.path('..', 'reports', subdir, paste0(stringr::str_pad(cnt, 3, pad = "0"),
                                                   '_pheatmap_logFC_num_', main, ".pdf"))
  
  # row_colors = cc[match(rownames(logFC.w), cc$display_name), ]$color_code

  ph = pheatmap::pheatmap(logFC.w, cluster_rows = FALSE, cluster_cols = FALSE,
                     color=myColor, breaks=myBreaks,
                     main = main,
                     display_numbers = TRUE,
                     fontsize_number = matrix(ifelse(p.w >= 0.05, 4, 8), nrow(p.w)), 
                     number_color = matrix(ifelse(p.w >= 0.05, "grey60", 'black'), nrow(p.w)),
                     filename = phfn)$gtable
  
  ggsave(filename = phfn,
         width = width, height = height, units = "in",
         plot=ph)
  

  dev.off()
      
  grid::grid.draw(ph)
  
    


}


```


# predefined contrasts



```{r, contrasts}

fpi = file.path('..', 'input')
fn = 'contrasts.txt'

comp = data.table::fread(file.path(fpi, fn), header = FALSE)


```


# metabolomics



```{r}

dir.create(file.path('..', 'reports', 'metabolomics'))

cnt = 1

```



```{r, metabolomics}

fpi = file.path('..', 'other')
fn = 'path.txt'
path = as.data.frame(data.table::fread(file.path(fpi, fn)))
path = path[order(path$pathway), ]

fpi = file.path('..', 'other')
fn = 'colourCodes_v2.txt'
cc = as.data.frame(data.table::fread(file.path(fpi, fn)))


fpi = file.path('..', 'input')


fn = "data_targeted.xlsx"
omics = openxlsx::read.xlsx(file.path(fpi, fn), sheet = 'metabolomics-leaves')
data.table::setDF(omics)

which(colnames(omics)[-1] %nin% cc$display_name)

Genotype = 'Desiree'
omics = cbind(Genotype, omics)


PostInductionDay = as.numeric(stringr::str_split_i(gsub('S', '', omics$SampleID), '_', i = 2))
omics = cbind(PostInductionDay, omics)

omics = omics[omics$PostInductionDay %in% 1:14, ]

Condition = gsub('_.*', '', omics$SampleID)
omics = cbind(Condition, omics)
omics = omics[-which(omics$Condition == 'W' & omics$PostInductionDay %in% c(8,14)), ]

omics$Condition = factor(omics$Condition, levels = c('C', 'H', 'D', 'HD', 'W'))

table(omics$Condition, omics$PostInductionDay)
              


mygroup = paste(omics$Condition, stringr::str_pad(omics$PostInductionDay, 1, pad = "0"), sep = '_')
# table(mygroup %in% comp$V1)
# table(mygroup %in% comp$V3)

omics = cbind(mygroup, omics)

```


## palette

```{r, metabolomics-palette}



n = 6


# my.ggplot.palette = brewer.pal(8, "Set2")[c(8, 2, 6, 4, 5, 3)]#[c(8, 2, 6, 7, 5, 3)]
my.ggplot.palette = c('#7F7F7F', '#C00000', '#FFC000', '#ED7D31',  
                      # '#70AD47', 
                      '#4472C4')
pie(rep(1, length(my.ggplot.palette)), 
    col = my.ggplot.palette, 
    main = 'ggplot palette',
    labels = c('C', 'H', 'D', 'HD', 
               # 'HDW', 
               'W'))

extend.palette = cbind(my.ggplot.palette, c('C', 'H', 'D', 'HD', 
                                            # 'HDW', 
                                            'W'))
colnames(extend.palette) = c('colour', 'group')
extend.palette = merge(extend.palette, omics[, 1:4], by.x = 'group', by.y = 'Condition', all = TRUE)


par(mfrow = c(2,2))

nn = 9
# (intervals = cut(c(-1,1), n*2))
intervals = cut(c(-1,1), 
                breaks = setdiff(seq(-1, 1, 0.1), 0), 
                include.lowest = TRUE)

my.heatmap.col1 = c(rev(brewer.pal(nn, 'Reds')), 'white', brewer.pal(nn, 'Blues'))
# my.heatmap.col1 = cividis(19, direction = -1)
pie(rep(1, length(my.heatmap.col1)), 
    col = rev(my.heatmap.col1), 
    labels = levels(intervals), 
    main = 'heatmap palette 1')


my.heatmap.col2 = c(my.heatmap.col1[1:3], 
                    rep('grey90', length(4:(nn*2-2))),
                    my.heatmap.col1[(nn*2-1): (nn*2 + 1)])
pie(rep(1, length(my.heatmap.col1)), 
    col = rev(my.heatmap.col2), 
    labels = levels(intervals), 
    main = 'heatmap palette 2')


par(mfrow = c(1,1))


```


## density

```{r, metabolomics-long}

# column in which measurements start
nx = 6

data = omics
# colnames(data)[nx:ncol(data)]

data[, nx:ncol(data)] = apply(data[, nx:ncol(data)] , 2, as.numeric)

# https://www.cookbook-r.com/Manipulating_data/Converting_data_between_wide_and_long_format/
data.long = tidyr::gather(data, molecule, measurement, colnames(data)[nx]:colnames(data)[ncol(data)], factor_key=FALSE)


data.long$measurement = as.numeric(data.long$measurement)

data.long$molecule = factor(data.long$molecule, levels = colnames(data)[nx:ncol(data)])

data.long$PostInductionDay = ordered(data.long$PostInductionDay, levels = sort(unique(data.long$PostInductionDay )))



data.long = data.long[!is.na(data.long$measurement), ]


# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = Condition, colour = Condition)) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = my.ggplot.palette) +
  scale_fill_manual(name = 'Legend',
                    values = my.ggplot.palette) 
# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = mygroup , colour = mygroup )) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = unique(extend.palette[, 2:3])[, 1]) +
  scale_fill_manual(name = 'Legend',
                    values = unique(extend.palette[, 2:3])[, 1]
                    ) 



data.long.SE = summarySE(data = data.long, 
                         measurevar="measurement", 
                         groupvars=c("molecule", "Genotype", "Condition", "PostInductionDay", "mygroup"), 
                         na.rm = TRUE)




```


```{r, metabolomics-ncol}

ncol = 6

jitter = position_jitter(width = 0.0, height = 0.0)
dodge = position_dodge(width=0.0)


```


## plot measurements

```{r, metabolomics-myggplot}

cd = sort(unique(data.long.SE$PostInductionDay))

mypallette = my.ggplot.palette

# useful in case of more genotypes
lims = data.long.SE %>% 
  dplyr::group_by(molecule) %>% 
  dplyr::summarise(Panel = molecule, 
                   ymin=min(mean-se),ymax=max(mean+se),
                   n = 5
                   )
lims = unique(lims)
lims = split(lims, lims$Panel) # make a list
lims = lapply(lims, function(l) {
  scale_y_continuous(limits = c(l$ymin, l$ymax),
                     n.breaks = l$n)
  }
)


title = 'Desiree'
data.SE = data.long.SE[data.long.SE$Genotype == title, ]
myggplot(data.SE, title, cnt, lncol = 2, lims, subdir = 'metabolomics')



  


```

## correlation

```{r, metabolomics-cor}

mytitle = 'Desiree'
df = omics[omics$Genotype == mytitle, ]
df = as.matrix(df[, nx:ncol(df)])
df = apply(df, 2, as.numeric)
cs = colSums(is.na(df))/nrow(df)
if (any(cs == 1)) df = df[, -which(cs == 1)]

rownames(df) = omics$SampleID[omics$Genotype == mytitle]

mycond = levels(omics$Condition)


# Open the PDF device
pdf(file.path('..', 'reports', 'metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")), 
    width = 13, height = 13)

# Create the plot
my.pairs.panels(df, mytitle, mycond, cccex = 1)

# Close the device to save the plot
dev.off()

# Now plot it to display in the document
my.pairs.panels(df, mytitle, mycond, cccex = 1)


my.heatmaply(df, mytitle, mycond, subdir = 'metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'metabolomics')





mycond = 'C'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'metabolomics')


mycond = 'H'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])



pdf(
         file.path('..', 'reports', 'metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'metabolomics')


mycond = 'D'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'metabolomics')


mycond = 'W'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'metabolomics')

    
mycond = 'HD'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'metabolomics')








```


```{r, metabolomics-stat}

genotype = 'Desiree'
De.log2FC = NULL
De.p = NULL

for (i in 1:nrow(comp)) {
  
  # print(i)
  # print( as.character(comp[i, ]))
  
  dtX = data.long[data.long$mygroup %in% comp[i, c(1,3)] & data.long$Genotype %in% genotype, ]

  stress = dtX[dtX$mygroup == as.character(comp[i, 1]), ]
  Ctrl = dtX[dtX$mygroup == as.character(comp[i, 3]), ]
  
  StressC = gsub('_.*', '', comp[i, 1])
  CtrlC = gsub('_.*', '', comp[i, 3])
  
  log2FC = my.logFC(Ctrl = Ctrl, stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  p = my.customised.t.test(Ctrl = Ctrl, Stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  
  De.log2FC = rbind(De.log2FC, log2FC)
  De.p = rbind(De.p, p)
  
  
}


Desiree.stat = merge(De.log2FC, De.p,
                     by.y = c("Genotype", "molecule", "group2", "group1"),
                     by.x = c("Genotype", "molecule", "stressGroup", "controlGroup"),
                     all = TRUE)
Desiree.stat$stressGroup = paste(gsub('_.*', '', Desiree.stat$stressGroup),
                                  stringr::str_pad(gsub('.*_', '', Desiree.stat$stressGroup), 2, pad = "0"),
                                  sep = '_')
logFC = tidyr::spread(Desiree.stat[, c(2:3, 7)], stressGroup, log2FC)
p = tidyr::spread(Desiree.stat[, c(2:3, 13)], stressGroup, p)
colnames(logFC)[2:ncol(logFC)] = paste(colnames(logFC)[2:ncol(logFC)], 'logFC', sep = ' | ')
colnames(p)[2:ncol(p)] = paste(colnames(p)[2:ncol(p)], 'p', sep = ' | ')
stat = merge(logFC, p, by = 'molecule')
ind = match(sort(colnames(stat)[2:ncol(stat)]), colnames(stat)[2:ncol(stat)])
stat = stat[, c(1, ind+1)]
stat[2:ncol(stat)] = format(round(as.data.frame(stat[2:ncol(stat)]), 3), nsmall = 3)
openxlsx::write.xlsx(stat, '../reports/Desiree.logFC-metabolomics_leaves.xlsx', asTable = TRUE)


```

## heatmap


```{r, metabolomics-heatmap}

comp = as.data.frame(comp)
forHeatmap = comp[intersect(grep('C', comp[, 1], invert = TRUE), grep('C', comp[, 3])), ]
forHeatmap[, 1] = sapply(forHeatmap[, 1], function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
forHeatmap[, 3] = sapply(forHeatmap[, 3], function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
forHeatmap$comparison = paste(forHeatmap[, 1], forHeatmap[, 3], sep = '-')
forHeatmap = forHeatmap[match(forHeatmap[, 1], c("H_01", "H_07","H_08","H_14",
                                                 "HD_08", "HD_14",
                                                 "D_08","D_14",
                                                 "W_01", "W_07")), ]

Desiree.stat$controlGroup = sapply(Desiree.stat$controlGroup, function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
df = Desiree.stat[Desiree.stat$stressGroup %in% forHeatmap[, 1] & Desiree.stat$controlGroup %in% forHeatmap[, 3], ]


my.pheatmap.logFC(df, forHeatmap,
                  main = 'Desiree-metabolomics',
                  cluster_rows = FALSE, cluster_cols = FALSE,
                  subdir = 'metabolomics',
                  width = 6, height = 6, cnt = cnt)





```




```{r, echo=TRUE, results='hide', warning=FALSE, message=FALSE}


metabolomics.leaves = stat[, grep('molecule|log', colnames(stat))]
# rm()
gc()
gc()

```



# hormonomics



```{r}

dir.create(file.path('..', 'reports', 'hormonomics'))

cnt = 2

```



```{r, hormonomics}

fpi = file.path('..', 'other')
fn = 'path.txt'
path = as.data.frame(data.table::fread(file.path(fpi, fn)))
path = path[order(path$pathway), ]

fpi = file.path('..', 'other')
fn = 'colourCodes_v2.txt'
cc = as.data.frame(data.table::fread(file.path(fpi, fn)))


fpi = file.path('..', 'input')


fn = "data_targeted.xlsx"
omics = openxlsx::read.xlsx(file.path(fpi, fn), sheet = 'hormonomics-leaves')
data.table::setDF(omics)

which(colnames(omics)[-1] %nin% cc$display_name)

Genotype = 'Desiree'
omics = cbind(Genotype, omics)


PostInductionDay = as.numeric(stringr::str_split_i(gsub('S|P', '', omics$SampleID), '_', i = 2))
omics = cbind(PostInductionDay, omics)

omics = omics[omics$PostInductionDay %in% 1:14, ]

Condition = gsub('_.*', '', omics$SampleID)
omics = cbind(Condition, omics)
omics = omics[-which(omics$Condition == 'W' & omics$PostInductionDay %in% c(8,14)), ]

omics$Condition = factor(omics$Condition, levels = c('C', 'H', 'D', 'HD', 'W'))

table(omics$Condition, omics$PostInductionDay)
              


mygroup = paste(omics$Condition, stringr::str_pad(omics$PostInductionDay, 1, pad = "0"), sep = '_')
# table(mygroup %in% comp$V1)
# table(mygroup %in% comp$V3)

omics = cbind(mygroup, omics)

```


## palette

```{r, hormonomics-palette}



my.ggplot.palette = c('#7F7F7F', '#C00000', '#FFC000', '#ED7D31',  
                      # '#70AD47', 
                      '#4472C4')

extend.palette = cbind(my.ggplot.palette, c('C', 'H', 'D', 'HD',
                                            #'HDW', 
                                            'W'))
colnames(extend.palette) = c('colour', 'group')
extend.palette = merge(extend.palette, omics[, 1:4], by.x = 'group', by.y = 'Condition', all = TRUE)




```


## density

```{r, hormonomics-long}

# column in which measurements start
nx = 6

data = omics
# colnames(data)[nx:ncol(data)]

data[, nx:ncol(data)] = apply(data[, nx:ncol(data)] , 2, as.numeric)

data.long = tidyr::gather(data, molecule, measurement, colnames(data)[nx]:colnames(data)[ncol(data)], factor_key=FALSE)


data.long$measurement = as.numeric(data.long$measurement)

data.long$molecule = factor(data.long$molecule, levels = colnames(data)[nx:ncol(data)])

data.long$PostInductionDay = ordered(data.long$PostInductionDay, levels = sort(unique(data.long$PostInductionDay )))


data.long = data.long[!is.na(data.long$measurement), ]


# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = Condition, colour = Condition)) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = my.ggplot.palette) +
  scale_fill_manual(name = 'Legend',
                    values = my.ggplot.palette) 
# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = mygroup , colour = mygroup )) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = unique(extend.palette[, 2:3])[, 1]) +
  scale_fill_manual(name = 'Legend',
                    values = unique(extend.palette[, 2:3])[, 1]
                    ) 



data.long.SE = summarySE(data = data.long, 
                         measurevar="measurement", 
                         groupvars=c("molecule", "Genotype", "Condition", "PostInductionDay", "mygroup"), 
                         na.rm = TRUE)




```



## plot measurements

```{r, hormonomics-myggplot}

cd = sort(unique(data.long.SE$PostInductionDay))

mypallette = my.ggplot.palette


lims = data.long.SE %>% 
  dplyr::group_by(molecule) %>% 
  dplyr::summarise(Panel = molecule, 
                   ymin=min(mean-se),ymax=max(mean+se),
                   n = 5
                   )
lims = unique(lims)
lims = split(lims, lims$Panel) # make a list
lims = lapply(lims, function(l) {
  scale_y_continuous(limits = c(l$ymin, l$ymax),
                     n.breaks = l$n)
  }
)


title = 'Desiree'
data.SE = data.long.SE[data.long.SE$Genotype == title, ]
myggplot(data.SE, title, cnt, lncol = 2, lims, subdir = 'hormonomics')



  


```


```{r}

omics  = omics[, -grep('neoPA', colnames(omics))]
data.long.SE  = data.long.SE[data.long.SE$molecule != 'neoPA', ]
data.long  = data.long[data.long$molecule != 'neoPA', ]


```

## correlation

```{r, hormonomics-cor}

mytitle = 'Desiree'
df = omics[omics$Genotype == mytitle, ]
df = as.matrix(df[, nx:ncol(df)])
df = apply(df, 2, as.numeric)
cs = colSums(is.na(df))/nrow(df)
if (any(cs == 1)) df = df[, -which(cs == 1)]

rownames(df) = omics$SampleID[omics$Genotype == mytitle]

mycond = levels(omics$Condition)


pdf(
         file.path('..', 'reports', 'hormonomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)


my.heatmaply(df, mytitle, mycond, subdir = 'hormonomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'hormonomics')





mycond = 'C'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'hormonomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'hormonomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'hormonomics')


mycond = 'H'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'hormonomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'hormonomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'hormonomics')


mycond = 'D'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'hormonomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'hormonomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'hormonomics')


mycond = 'W'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'hormonomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'hormonomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'hormonomics')

    
mycond = 'HD'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'hormonomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'hormonomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'hormonomics')








```


```{r, hormonomics-stat}

genotype = 'Desiree'
De.log2FC = NULL
De.p = NULL

for (i in 1:nrow(comp)) {
  
  # print(i)
  # print( as.character(comp[i, ]))
  
  dtX = data.long[data.long$mygroup %in% comp[i, c(1,3)] & data.long$Genotype %in% genotype, ]

  stress = dtX[dtX$mygroup == as.character(comp[i, 1]), ]
  Ctrl = dtX[dtX$mygroup == as.character(comp[i, 3]), ]
  
  StressC = gsub('_.*', '', comp[i, 1])
  CtrlC = gsub('_.*', '', comp[i, 3])
  
  log2FC = my.logFC(Ctrl = Ctrl, stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  p = my.customised.t.test(Ctrl = Ctrl, Stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  
  De.log2FC = rbind(De.log2FC, log2FC)
  De.p = rbind(De.p, p)
  
  
}


Desiree.stat = merge(De.log2FC, De.p,
                     by.y = c("Genotype", "molecule", "group2", "group1"),
                     by.x = c("Genotype", "molecule", "stressGroup", "controlGroup"),
                     all = TRUE)
Desiree.stat$stressGroup = paste(gsub('_.*', '', Desiree.stat$stressGroup),
                                  stringr::str_pad(gsub('.*_', '', Desiree.stat$stressGroup), 2, pad = "0"),
                                  sep = '_')
logFC = tidyr::spread(Desiree.stat[, c(2:3, 7)], stressGroup, log2FC)
p = tidyr::spread(Desiree.stat[, c(2:3, 13)], stressGroup, p)
colnames(logFC)[2:ncol(logFC)] = paste(colnames(logFC)[2:ncol(logFC)], 'logFC', sep = ' | ')
colnames(p)[2:ncol(p)] = paste(colnames(p)[2:ncol(p)], 'p', sep = ' | ')
stat = merge(logFC, p, by = 'molecule')
ind = match(sort(colnames(stat)[2:ncol(stat)]), colnames(stat)[2:ncol(stat)])
stat = stat[, c(1, ind+1)]
stat[2:ncol(stat)] = format(round(as.data.frame(stat[2:ncol(stat)]), 3), nsmall = 3)
openxlsx::write.xlsx(stat, '../reports/Desiree.logFC-hormonomics_leaves.xlsx', asTable = TRUE)

```

## heatmap


```{r, hormonomics-heatmap}

comp = as.data.frame(comp)
forHeatmap = comp[intersect(grep('C', comp[, 1], invert = TRUE), grep('C', comp[, 3])), ]
forHeatmap[, 1] = sapply(forHeatmap[, 1], function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
forHeatmap[, 3] = sapply(forHeatmap[, 3], function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
forHeatmap$comparison = paste(forHeatmap[, 1], forHeatmap[, 3], sep = '-')
forHeatmap = forHeatmap[match(forHeatmap[, 1], c("H_01", "H_07","H_08","H_14",
                                                 "HD_08", "HD_14",
                                                 "D_08","D_14",
                                                 "W_01", "W_07")), ]

Desiree.stat$controlGroup = sapply(Desiree.stat$controlGroup, function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
df = Desiree.stat[Desiree.stat$stressGroup %in% forHeatmap[, 1] & Desiree.stat$controlGroup %in% forHeatmap[, 3], ]


my.pheatmap.logFC(df, forHeatmap,
                  main = 'Desiree-hormonomics',
                  cluster_rows = FALSE, cluster_cols = FALSE,
                  subdir = 'hormonomics',
                  width = 6, height = 6, cnt = cnt)





```



```{r, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

hormonomics.leaves = stat[, grep('molecule|log', colnames(stat))]

# rm()
gc()
gc()

```


# transcriptomics



```{r}

dir.create(file.path('..', 'reports', 'transcriptomics'))

cnt = 3

```



```{r, transcriptomics}

fpi = file.path('..', 'other')
fn = 'path.txt'
path = as.data.frame(data.table::fread(file.path(fpi, fn)))
path = path[order(path$pathway), ]

fpi = file.path('..', 'other')
fn = 'colourCodes_v2.txt'
cc = as.data.frame(data.table::fread(file.path(fpi, fn)))



fpi = file.path('..', 'input')


fn = "data_targeted.xlsx"
omics = openxlsx::read.xlsx(file.path(fpi, fn), sheet = 'transcriptomics-leaves')
data.table::setDF(omics)


which(colnames(omics)[-1] %nin% cc$display_name)
colnames(omics)[grep('RD29B', colnames(omics))] = 'RD29'
colnames(omics)[grep('M0ZJG3', colnames(omics))] = 'MES'

Genotype = 'Desiree'
omics = cbind(Genotype, omics)


PostInductionDay = as.numeric(stringr::str_split_i(gsub('S|P', '', omics$SampleID), '_', i = 2))
omics = cbind(PostInductionDay, omics)

omics = omics[omics$PostInductionDay %in% 1:14, ]

Condition = gsub('_.*', '', omics$SampleID)
omics = cbind(Condition, omics)
omics = omics[-which(omics$Condition == 'W' & omics$PostInductionDay %in% c(8,14)), ]

omics$Condition = factor(omics$Condition, levels = c('C', 'H', 'D', 'HD', 'W'))

table(omics$Condition, omics$PostInductionDay)
              


mygroup = paste(omics$Condition, stringr::str_pad(omics$PostInductionDay, 1, pad = "0"), sep = '_')
# table(mygroup %in% comp$V1)
# table(mygroup %in% comp$V3)

omics = cbind(mygroup, omics)

```


## palette

```{r, transcriptomics-palette}



my.ggplot.palette = c('#7F7F7F', '#C00000', '#FFC000', '#ED7D31',  
                      # '#70AD47', 
                      '#4472C4')

extend.palette = cbind(my.ggplot.palette, c('C', 'H', 'D', 'HD',
                                            #'HDW', 
                                            'W'))
colnames(extend.palette) = c('colour', 'group')
extend.palette = merge(extend.palette, omics[, 1:4], by.x = 'group', by.y = 'Condition', all = TRUE)




```


## density

```{r, transcriptomics-long}

# column in which measurements start
nx = 6

data = omics
# colnames(data)[nx:ncol(data)]

data[, nx:ncol(data)] = apply(data[, nx:ncol(data)] , 2, as.numeric)

data.long = tidyr::gather(data, molecule, measurement, colnames(data)[nx]:colnames(data)[ncol(data)], factor_key=FALSE)


data.long$measurement = as.numeric(data.long$measurement)

data.long$molecule = factor(data.long$molecule, levels = colnames(data)[nx:ncol(data)])

data.long$PostInductionDay = ordered(data.long$PostInductionDay, levels = sort(unique(data.long$PostInductionDay )))


data.long = data.long[!is.na(data.long$measurement), ]


# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = Condition, colour = Condition)) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = my.ggplot.palette) +
  scale_fill_manual(name = 'Legend',
                    values = my.ggplot.palette) 
# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = mygroup , colour = mygroup )) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = unique(extend.palette[, 2:3])[, 1]) +
  scale_fill_manual(name = 'Legend',
                    values = unique(extend.palette[, 2:3])[, 1]
                    ) 



data.long.SE = summarySE(data = data.long, 
                         measurevar="measurement", 
                         groupvars=c("molecule", "Genotype", "Condition", "PostInductionDay", "mygroup"), 
                         na.rm = TRUE)




```



## plot measurements

```{r, transcriptomics-myggplot}

cd = sort(unique(data.long.SE$PostInductionDay))

mypallette = my.ggplot.palette


lims = data.long.SE %>% 
  dplyr::group_by(molecule) %>% 
  dplyr::summarise(Panel = molecule, 
                   ymin=min(mean-se),ymax=max(mean+se),
                   n = 5
                   )
lims = unique(lims)
lims = split(lims, lims$Panel) # make a list
lims = lapply(lims, function(l) {
  scale_y_continuous(limits = c(l$ymin, l$ymax),
                     n.breaks = l$n)
  }
)


title = 'Desiree'
data.SE = data.long.SE[data.long.SE$Genotype == title, ]
myggplot(data.SE, title, cnt, lncol = 2, lims, subdir = 'transcriptomics')



  


```


## correlation

```{r, transcriptomics-cor}

mytitle = 'Desiree'
df = omics[omics$Genotype == mytitle, ]
df = as.matrix(df[, nx:ncol(df)])
df = apply(df, 2, as.numeric)
cs = colSums(is.na(df))/nrow(df)
if (any(cs == 1)) df = df[, -which(cs == 1)]

rownames(df) = omics$SampleID[omics$Genotype == mytitle]

mycond = levels(omics$Condition)


pdf(
         file.path('..', 'reports', 'transcriptomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)


my.heatmaply(df, mytitle, mycond, subdir = 'transcriptomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'transcriptomics')





mycond = 'C'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'transcriptomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'transcriptomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'transcriptomics')


mycond = 'H'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'transcriptomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'transcriptomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'transcriptomics')


mycond = 'D'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'transcriptomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'transcriptomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'transcriptomics')


mycond = 'W'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'transcriptomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'transcriptomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'transcriptomics')

    
mycond = 'HD'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'transcriptomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'transcriptomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'transcriptomics')








```


```{r, transcriptomics-stat}

genotype = 'Desiree'
De.log2FC = NULL
De.p = NULL

for (i in 1:nrow(comp)) {
  
  # print(i)
  # print( as.character(comp[i, ]))
  
  dtX = data.long[data.long$mygroup %in% comp[i, c(1,3)] & data.long$Genotype %in% genotype, ]

  stress = dtX[dtX$mygroup == as.character(comp[i, 1]), ]
  Ctrl = dtX[dtX$mygroup == as.character(comp[i, 3]), ]
  
  StressC = gsub('_.*', '', comp[i, 1])
  CtrlC = gsub('_.*', '', comp[i, 3])
  
  log2FC = my.logFC(Ctrl = Ctrl, stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  p = my.customised.t.test(Ctrl = Ctrl, Stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  
  De.log2FC = rbind(De.log2FC, log2FC)
  De.p = rbind(De.p, p)
  
  
}


Desiree.stat = merge(De.log2FC, De.p,
                     by.y = c("Genotype", "molecule", "group2", "group1"),
                     by.x = c("Genotype", "molecule", "stressGroup", "controlGroup"),
                     all = TRUE)
Desiree.stat$stressGroup = paste(gsub('_.*', '', Desiree.stat$stressGroup),
                                  stringr::str_pad(gsub('.*_', '', Desiree.stat$stressGroup), 2, pad = "0"),
                                  sep = '_')
logFC = tidyr::spread(Desiree.stat[, c(2:3, 7)], stressGroup, log2FC)
p = tidyr::spread(Desiree.stat[, c(2:3, 13)], stressGroup, p)
colnames(logFC)[2:ncol(logFC)] = paste(colnames(logFC)[2:ncol(logFC)], 'logFC', sep = ' | ')
colnames(p)[2:ncol(p)] = paste(colnames(p)[2:ncol(p)], 'p', sep = ' | ')
stat = merge(logFC, p, by = 'molecule')
ind = match(sort(colnames(stat)[2:ncol(stat)]), colnames(stat)[2:ncol(stat)])
stat = stat[, c(1, ind+1)]
stat[2:ncol(stat)] = format(round(as.data.frame(stat[2:ncol(stat)]), 3), nsmall = 3)
openxlsx::write.xlsx(stat, '../reports/Desiree.logFC-transcriptomics_leaves.xlsx', asTable = TRUE)

```

## heatmap


```{r, transcriptomics-heatmap}

comp = as.data.frame(comp)
forHeatmap = comp[intersect(grep('C', comp[, 1], invert = TRUE), grep('C', comp[, 3])), ]
forHeatmap[, 1] = sapply(forHeatmap[, 1], function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
forHeatmap[, 3] = sapply(forHeatmap[, 3], function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
forHeatmap$comparison = paste(forHeatmap[, 1], forHeatmap[, 3], sep = '-')
forHeatmap = forHeatmap[match(forHeatmap[, 1], c("H_01", "H_07","H_08","H_14",
                                                 "HD_08", "HD_14",
                                                 "D_08","D_14",
                                                 "W_01", "W_07")), ]

Desiree.stat$controlGroup = sapply(Desiree.stat$controlGroup, function(x) paste0(gsub("_(\\d+)", '', x), sprintf("_%02d", as.numeric(gsub(".*_", "", x)))))
df = Desiree.stat[Desiree.stat$stressGroup %in% forHeatmap[, 1] & Desiree.stat$controlGroup %in% forHeatmap[, 3], ]


my.pheatmap.logFC(df
                  , forHeatmap
                  , main = 'Desiree-transcriptomics'
                  , cluster_rows = FALSE
                  , cluster_cols = FALSE
                  , subdir = 'transcriptomics'
                  , width = 6
                  , height = 6
                  , cnt = cnt
                  )





```



```{r, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

transcriptomics.leaves = stat[, grep('molecule|log', colnames(stat))]

# rm()
gc()
gc()

```




# tubers metabolomics



```{r}

dir.create(file.path('..', 'reports', 'tubers-metabolomics'))

cnt = 4

```



```{r, tubers-metabolomics}

fpi = file.path('..', 'other')
fn = 'path.txt'
path = as.data.frame(data.table::fread(file.path(fpi, fn)))
path = path[order(path$pathway), ]

fpi = file.path('..', 'other')
fn = 'colourCodes_v2.txt'
cc = as.data.frame(data.table::fread(file.path(fpi, fn)))

fpi = file.path('..', 'input')


fn = "data_targeted.xlsx"
omics = openxlsx::read.xlsx(file.path(fpi, fn), sheet = 'metabolomics-tubers')
data.table::setDF(omics)

which(colnames(omics)[-c(1:4)] %nin% cc$display_name)

Genotype = 'Desiree'
omics = cbind(Genotype, omics)


PostInductionDay = as.numeric(stringr::str_split_i(gsub('S|P', '', omics$SampleID), '_', i = 2))
omics = cbind(PostInductionDay, omics)


Condition = gsub('_.*', '', omics$SampleID)
omics = cbind(Condition, omics)

omics$Condition = factor(omics$Condition, levels = c('C', 'H', 'D', 'HD', 'W'))

table(omics$Condition, omics$PostInductionDay)
              


mygroup = paste(omics$Condition, stringr::str_pad(omics$PostInductionDay, 1, pad = "0"), sep = '_')
# table(mygroup %in% comp$V1)
# table(mygroup %in% comp$V3)

omics = cbind(mygroup, omics)

```


## palette

```{r, tubers-metabolomics-palette}



my.ggplot.palette = c('#7F7F7F', '#C00000', '#FFC000', '#ED7D31',  
                      # '#70AD47', 
                      '#4472C4')

extend.palette = cbind(my.ggplot.palette, c('C', 'H', 'D', 'HD',
                                            #'HDW', 
                                            'W'))
colnames(extend.palette) = c('colour', 'group')
extend.palette = merge(extend.palette, omics[, 1:4], by.x = 'group', by.y = 'Condition', all = TRUE)




```


## density

```{r, tubers-metabolomics-long}

# column in which measurements start
nx = 9

data = omics
# colnames(data)[nx:ncol(data)]

data[, nx:ncol(data)] = apply(data[, nx:ncol(data)] , 2, as.numeric)

data.long = tidyr::gather(data, molecule, measurement, colnames(data)[nx]:colnames(data)[ncol(data)], factor_key=FALSE)


data.long$measurement = as.numeric(data.long$measurement)

data.long$molecule = factor(data.long$molecule, levels = colnames(data)[nx:ncol(data)])

data.long$PostInductionDay = ordered(data.long$PostInductionDay, levels = sort(unique(data.long$PostInductionDay )))


data.long = data.long[!is.na(data.long$measurement), ]


# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = Condition, colour = Condition)) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = my.ggplot.palette) +
  scale_fill_manual(name = 'Legend',
                    values = my.ggplot.palette) 
# X11()
ggplot(data.long[data.long$Genotype == 'Desiree', ], aes(log2(measurement), fill = mygroup , colour = mygroup )) +
  geom_density(alpha = 0.1) + 
  ggtitle('Desiree')  +
  theme_classic() +
  scale_colour_manual(name = 'Legend',
                      values = unique(extend.palette[, 2:3])[, 1]) +
  scale_fill_manual(name = 'Legend',
                    values = unique(extend.palette[, 2:3])[, 1]
                    ) 



data.long.SE = summarySE(data = data.long, 
                         measurevar="measurement", 
                         groupvars=c("molecule", "Genotype", "Condition", "PostInductionDay", "mygroup"), 
                         na.rm = TRUE)




```



## plot measurements

```{r, tubers-metabolomics-myggplot}

cd = sort(unique(data.long.SE$PostInductionDay))

mypallette = my.ggplot.palette




title = 'Tubers'
lncol = 2
subdir = 'tubers-metabolomics'

# X11();
myplotY = ggplot(data.long, aes(x = Condition , 
                                y = measurement, 
                                group = Condition,
                                shape = Condition,
                                fill = Condition,
                                colour = Condition)) + 
  geom_boxplot() + 
      facet_wrap(~ molecule, ncol = ncol,
             strip.position = "top",
             scales = 'free_y',
             drop = FALSE
             ) +
      ggtitle(title)  +
    theme_classic() +
      scale_colour_manual(name = 'Legend',
                          values = mypallette) +
      scale_fill_manual(name = 'Legend',
                        values = mypallette
                        ) + 
        labs(x = '', 
             y = 'Relative molecule abundance [+/- SE]') + 
        scale_y_continuous(# trans = log2_trans(), 
                           labels = scales::comma
                           ) +

      theme(axis.text.x = element_text (angle = 0, # 45, 
                                        vjust = 1, 
                                        hjust = 1, 
                                        size = 12),
            axis.text.y = element_text (angle = 0, 
                                        vjust = 1, 
                                        hjust = 1, 
                                        size = 12),
            axis.title.x = element_text (angle = 0, # 45, 
                                        #vjust = 1, 
                                        #hjust = 1, 
                                        size = 12,
                                        face = "bold"),
            axis.title.y = element_text (angle = 90, 
                                        # vjust = 1, 
                                        # hjust = 1, 
                                        size = 12,
                                        face = "bold"),
            axis.text = element_text( angle = 90 ),
            panel.grid.major = element_blank(),
            panel.grid.minor = element_blank(),
            panel.border = element_blank(),
            panel.background = element_blank(),
            legend.position = "bottom",
            legend.background = element_rect(fill=NULL, 
                                             size=0.1, linetype="solid"),
            legend.title = element_text(color = NULL,
                                        size = 12,
                                        face = "bold"),
            legend.text = element_text( size = 12),
            strip.text = element_text(size = 12)) + 
  theme(
        legend.box = "horizontal",
        legend.key.width = unit(2, 'cm'),
        legend.position = c(1, 0),
        legend.justification = c(1, 0)) +
  guides(colour = guide_legend(ncol = lncol)) +
  guides(linetype = guide_legend(ncol = lncol)) +
  guides(shape = "none") +
  theme(panel.spacing.y = unit(1.5, "lines")) +
  labs(color  = "Legend", linetype = "Legend", shape = "Legend")

plot(myplotY)  
# dev.off()
  
  
  fng = paste0(stringr::str_pad(cnt, 3, pad = "0"), '_', title, "_L.pdf")
  ggsave(file.path('..', 'reports', subdir, fng), 
         width = 16, height = 12, units = "in")
  fng = paste0(stringr::str_pad(cnt, 3, pad = "0"), '_', title, "_L.svg")
  ggsave(file.path('..', 'reports', subdir, fng),
         width = 16, height = 12, units = "in")
  


```


## correlation

```{r, tubers-metabolomics-cor}

mytitle = 'Desiree'
df = omics[omics$Genotype == mytitle, ]
df = as.matrix(df[, nx:ncol(df)])
df = apply(df, 2, as.numeric)
cs = colSums(is.na(df))/nrow(df)
if (any(cs == 1)) df = df[, -which(cs == 1)]

rownames(df) = omics$SampleID[omics$Genotype == mytitle]

mycond = levels(omics$Condition)


pdf(
         file.path('..', 'reports', 'tubers-metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)


my.heatmaply(df, mytitle, mycond, subdir = 'tubers-metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'tubers-metabolomics')





mycond = 'C'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'tubers-metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'tubers-metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'tubers-metabolomics')


mycond = 'H'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'tubers-metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'tubers-metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'tubers-metabolomics')


mycond = 'D'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'tubers-metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'tubers-metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'tubers-metabolomics')


mycond = 'W'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'tubers-metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'tubers-metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'tubers-metabolomics')

    
mycond = 'HD'    
df = omics[omics$Genotype == mytitle & omics$Condition == mycond, ]
df = as.matrix(df[, nx:ncol(df)])


pdf(
         file.path('..', 'reports', 'tubers-metabolomics', paste0('SPLOM_', 'PCC_', mytitle, '_', paste(mycond, collapse = '-'), ".pdf")),
         width = 13, height = 13)
my.pairs.panels(df, mytitle, mycond, cccex = 1)
dev.off()
my.pairs.panels(df, mytitle, mycond, cccex = 1)

my.heatmaply(df, mytitle, mycond, subdir = 'tubers-metabolomics') 

my.pheatmap(df, mytitle, mycond, subdir = 'tubers-metabolomics')








```


```{r, tubers-metabolomics-stat}

genotype = 'Desiree'
De.log2FC = NULL
De.p = NULL

comp = rbind(cbind('H_28',  '-', 'C_28'),
             cbind('D_28',  '-', 'C_28'),
             cbind('HD_28',  '-', 'C_28'),
             cbind('W_28',  '-', 'C_28'))

for (i in 1:nrow(comp)) {
  
  # print(i)
  # print( as.character(comp[i, ]))
  
  dtX = data.long[data.long$mygroup %in% comp[i, c(1,3)] & data.long$Genotype %in% genotype, ]

  stress = dtX[dtX$mygroup == as.character(comp[i, 1]), ]
  Ctrl = dtX[dtX$mygroup == as.character(comp[i, 3]), ]
  
  StressC = gsub('_.*', '', comp[i, 1])
  CtrlC = gsub('_.*', '', comp[i, 3])
  
  log2FC = my.logFC(Ctrl = Ctrl, stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  p = my.customised.t.test(Ctrl = Ctrl, Stress = stress, StressC = StressC, CtrlC = CtrlC, genotype = genotype)
  
  De.log2FC = rbind(De.log2FC, log2FC)
  De.p = rbind(De.p, p)
  
  
}


Desiree.stat = merge(De.log2FC, De.p,
                     by.y = c("Genotype", "molecule", "group2", "group1"),
                     by.x = c("Genotype", "molecule", "stressGroup", "controlGroup"),
                     all = TRUE)
Desiree.stat$stressGroup = paste(gsub('_.*', '', Desiree.stat$stressGroup),
                                  stringr::str_pad(gsub('.*_', '', Desiree.stat$stressGroup), 2, pad = "0"),
                                  sep = '_')
logFC = tidyr::spread(Desiree.stat[, c(2:3, 7)], stressGroup, log2FC)
p = tidyr::spread(Desiree.stat[, c(2:3, 13)], stressGroup, p)
colnames(logFC)[2:ncol(logFC)] = paste(colnames(logFC)[2:ncol(logFC)], 'logFC', sep = ' | ')
colnames(p)[2:ncol(p)] = paste(colnames(p)[2:ncol(p)], 'p', sep = ' | ')
stat = merge(logFC, p, by = 'molecule')
ind = match(sort(colnames(stat)[2:ncol(stat)]), colnames(stat)[2:ncol(stat)])
stat = stat[, c(1, ind+1)]
stat[2:ncol(stat)] = format(round(as.data.frame(stat[2:ncol(stat)]), 3), nsmall = 3)
openxlsx::write.xlsx(stat, '../reports/Desiree.logFC-metabolomics_tubers.xlsx', asTable = TRUE)

```

## heatmap


```{r, tubers-metabolomics-heatmap}

comp = as.data.frame(comp)
forHeatmap = comp[intersect(grep('C', comp[, 1], invert = TRUE), grep('C', comp[, 3])), ]
forHeatmap$comparison = paste(forHeatmap[, 1], forHeatmap[, 3], sep = '-')

df = Desiree.stat[Desiree.stat$stressGroup %in% forHeatmap[, 1] & Desiree.stat$controlGroup %in% forHeatmap[, 3], ]

my.pheatmap.logFC(df, forHeatmap,
                  main = 'Desiree-tubers-metabolomics',
                  cluster_rows = FALSE, cluster_cols = FALSE,
                  subdir = 'tubers-metabolomics',
                  width = 6, height = 6, cnt = cnt)





```



```{r, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

metabolomics.tubers = stat[, grep('molecule|log', colnames(stat))]

# rm()
gc()
gc()

```

# logFC results

```{r, echo=TRUE, warning=FALSE, message=FALSE}


targeted = rbind(metabolomics.leaves, hormonomics.leaves, transcriptomics.leaves)
targeted = merge(targeted, metabolomics.tubers, by = 'molecule', all.x = TRUE, all.y = TRUE)

data.table::fwrite(targeted, file.path('..', 'output', 'targeted_logFC.txt'), sep = '\t')

```


# Session info

```{r, session_info}

# devtools::session_info()
sessionInfo()


```
