==> _p_ADAPTOmics/_I_Desiree/_S_multiOmics/_A_multiOmicsStat-R/scripts/04.2_targeted-permute.Rmd <==
---
title: "04.2_targeted-permute"
author: "zagor"
date: "`r Sys.Date()`"
output:
  html_document:
    fig_caption: yes
    self_contained: yes
    fig_width: 9
    fig_height: 6
    toc: true
    toc_float:
      toc_collapsed: false
    toc_depth: 5
    number_sections: yes
    theme: flatly
    highlight: tango
  pdf_document:
    toc: yes
    toc_depth: '5'
  word_document:
    toc: yes
    toc_depth: '5'
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(#dev = c('pdf', 'png'),  # this embeds pdf and crates scrolable blocks
                      dev = c('png'), 
                      fig.align = 'center', 
                      fig.height = 6, 
                      fig.width = 9 ,
                      warning = FALSE, message = FALSE
                      )

```



```{r, rm, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

rm(list = ls(all = TRUE))

gc()
gc()

seed = 123456
set.seed(seed)

```


# libraries

```{r, libraries, echo=TRUE, warning=FALSE, message=FALSE}



# `%!in%` = Negate(`%in%`)
`%nin%` = Negate(`%in%`)

library(magrittr)
library(MKinfer)


# remotes::install_github("IndrajeetPatil/broomExtra")


```


# leaves

## data

```{r, omics, echo=TRUE, warning=FALSE, message=FALSE}


fp = file.path('..', 'input')
fn = 'data_targeted.xlsx'

horm = openxlsx::read.xlsx(xlsxFile = file.path(fp, fn),
                            sheet = 'hormonomics-leaves',
                            startRow = 1,
                            colNames = TRUE,
                            rowNames = FALSE,
                            detectDates = FALSE,
                            skipEmptyRows = TRUE,
                            skipEmptyCols = TRUE,
                            rows = NULL,
                            cols = NULL,
                            check.names = FALSE,
                            sep.names = ".",
                            namedRegion = NULL,
                            na.strings = "NA",
                            fillMergedCells = FALSE)
data.table::setDF(horm)
horm$Stress = gsub('_.*', '', horm$SampleID)
horm$Time = as.numeric(gsub('[SP]', '', stringr::str_split_i(horm$SampleID, "_", i = 2)))
horm$Plant = gsub('.*_', '', horm$SampleID)

met = openxlsx::read.xlsx(xlsxFile = file.path(fp, fn),
                            sheet = 'metabolomics-leaves',
                            startRow = 1,
                            colNames = TRUE,
                            rowNames = FALSE,
                            detectDates = FALSE,
                            skipEmptyRows = TRUE,
                            skipEmptyCols = TRUE,
                            rows = NULL,
                            cols = NULL,
                            check.names = FALSE,
                            sep.names = ".",
                            namedRegion = NULL,
                            na.strings = "NA",
                            fillMergedCells = FALSE)
data.table::setDF(met)
met$Stress = gsub('_.*', '', met$SampleID)
met$Time = as.numeric(gsub('[SP]', '', stringr::str_split_i(met$SampleID, "_", i = 2)))
met$Plant = gsub('.*_', '', met$SampleID)

qPCR = openxlsx::read.xlsx(xlsxFile = file.path(fp, fn),
                            sheet = 'transcriptomics-leaves',
                            startRow = 1,
                            colNames = TRUE,
                            rowNames = FALSE,
                            detectDates = FALSE,
                            skipEmptyRows = TRUE,
                            skipEmptyCols = TRUE,
                            rows = NULL,
                            cols = NULL,
                            check.names = FALSE,
                            sep.names = ".",
                            namedRegion = NULL,
                            na.strings = "NA",
                            fillMergedCells = FALSE)
data.table::setDF(qPCR)
qPCR$Stress = gsub('_.*', '', qPCR$SampleID)
qPCR$Time = as.numeric(gsub('[SP]', '', stringr::str_split_i(qPCR$SampleID, "_", i = 2)))
qPCR$Plant = gsub('.*_', '', qPCR$SampleID)

horm = horm[, -grep("neoPA|SampleID", colnames(horm))]
met = met[, -grep("SampleID", colnames(met))]
qPCR = qPCR[, -grep("SampleID", colnames(qPCR))]


omics = merge (met, horm, by = c("Stress", "Time", "Plant"), all = TRUE)
omics = merge (omics, qPCR, by = c("Stress", "Time", "Plant"), all = TRUE)


omics = omics[(omics$Stress %in% c('C', 'H') & omics$Time %in% 1:14) |
                (omics$Stress %in% c('D', 'HD') & omics$Time %in% 8:14) |
                (omics$Stress %in% c('W') & omics$Time %in% 1:7) , ]




```



```{r, DTC, echo=TRUE, warning=FALSE, message=FALSE}

stat.test = NULL
# typeof(omics$Time)


omics = omics[order(omics$Stress, omics$Time, omics$Plant), ]
rownames(omics) = paste(omics[, 1], omics[, 2], omics[, 3], sep = '_')


DTC = omics[, -grep('Plant', colnames(omics))]
DTC$Stress = factor(DTC$Stress, levels = c("C",   "D",   "H",   "HD",  "W"))
DTC$Time = factor(DTC$Time, levels = c(1, 7, 8, 14))
DTC = tidyr::gather(DTC, variable, measurement, colnames(DTC)[3]:colnames(DTC)[ncol(DTC)], factor_key=TRUE)
DTC = DTC[!is.na(DTC$measurement), ]
DTC = gdata::drop.levels(DTC)
DTC$variable = factor(DTC$variable)
DTC[DTC$measurement == 0, ]$measurement = NA # LOD

```



## add noise to avoid ties

```{r, noise, echo=TRUE, warning=FALSE, message=FALSE}


# identical values
remove = c(which(DTC$variable %in% c('RD29B', 'SP6A') & DTC$Time %in% 1:8 & DTC$Stress %in% 'H'))
DTC = DTC[-remove, ]
fp = file.path('..', 'other')
# added noise at the last decimal
fn = 'replace-with-jitter.txt'
replace = data.table::fread(file.path(fp, fn), header = FALSE)
data.table::setDF(replace)
colnames(replace) = colnames(DTC)
DTC = rbind(DTC, replace)


```


## stat test

```{r, stat, echo=TRUE, warning=FALSE, message=FALSE}


# en = levels(DTC$variable)[1]
# eff = 'D'
# j = 8

stat.test = NULL


for (en in levels(DTC$variable)) {


  set.seed(seed)

  el = NULL
  el = levels(DTC$Stress)[-1]


  for (eff in el) {


    # cat(en, '\t', eff, '\n')






  mydata.long = NULL
        mydata.long = dplyr::as_tibble(data.table::data.table(DTC[DTC$variable == en &
                                                                   DTC$Stress %in% c('C', eff), ]))
        
        

        if (eff %in% c('D', 'HD')) {
          mydata.long = mydata.long[mydata.long$Time %in% 8:14, ]
        }
        if (eff %in% c('W')) {
          mydata.long = mydata.long[mydata.long$Time %in% 1:7, ]
        }

        mydata.long = gdata::drop.levels(mydata.long)


          mytable = NULL

            for (j in levels(mydata.long$Time)) {
              
              # print(j)

                a = mydata.long[mydata.long$Time == j , ]
                a = gdata::drop.levels(a)
                
                if (length(levels(a$Stress)) > 1) {
                  
                  model = MKinfer::perm.t.test(log2(measurement) ~ Stress ,
                                                   data=a,
                                                   alternative = "two.sided",
                                                   mu = 0,
                                                   paired = FALSE,
                                                   var.equal = FALSE,
                                                   conf.level = 0.95,
                                                   symmetric = TRUE, # a logical variable indicating whether to assume symmetry in the two-sided test. If TRUE then the symmetric permutation p value otherwise the equal-tail permutation p value is computed.
                                                   p.adjust.method = "none",
                                                   detailed = TRUE,
                                                   R = 999, set.seed(123456),
                                              simplify = TRUE)
                  myvec = c(en, eff, j, model$method, model$p.value, model$perm.p.value)
                  mytable = rbind(mytable, myvec)
                }

              
            
          }
          mytable = as.data.frame(mytable)
          colnames(mytable) = c('Variable', 'Treatment', 'Time', 'Method', 'WelchP', 'PermP')


          stat.test = rbind(stat.test, mytable)



  }



} 



```

## wide format

```{r, wide, echo=TRUE, warning=FALSE, message=FALSE}


stat.test$Treatment = paste0(stat.test$Treatment, 
                                 formatC(as.numeric(as.character(stat.test$Time)), width=2, flag="0"))
ind = grep('Time', colnames(stat.test))
stat.test = stat.test[, -ind]
stat.test$WelchP = format(round(as.numeric(stat.test$WelchP), 3), nsmall = 3)
stat.test$PermP = format(round(as.numeric(stat.test$PermP), 3), nsmall = 3)


stat.test.wide2 = tidyr::spread(stat.test[, c(1:2, 5)], Treatment, PermP)
stat.test.wide1 = tidyr::spread(stat.test[, c(1:2, 4)], Treatment, WelchP)

colnames(stat.test.wide2)[2:ncol(stat.test.wide2)] = paste(colnames(stat.test.wide2)[2:ncol(stat.test.wide2)], 
                                                                   'PermP', sep = ' | ')
colnames(stat.test.wide1)[2:ncol(stat.test.wide1)] = paste(colnames(stat.test.wide1)[2:ncol(stat.test.wide1)], 
                                                                   'WelchP', sep = ' | ')


stat.test.wide = merge(stat.test.wide1, stat.test.wide2, by = c('Variable'), all = TRUE)
ind = match(sort(colnames(stat.test.wide)[2:ncol(stat.test.wide)]), colnames(stat.test.wide)[2:ncol(stat.test.wide)])
stat.test.wide = stat.test.wide[, c(1, ind+1)]

```

## write

```{r, res, echo=TRUE, warning=FALSE, message=FALSE}



# n = stat.test.wide$Variable
# 
# stat.test.wide = stat.test.wide
# ind = which(stat.test.wide < 0.05, arr.ind = TRUE)
# stat.test.wide[ind] = '*'
# ind = which(stat.test.wide >= 0.05, arr.ind = TRUE)
# stat.test.wide[ind] = 'ns'
# stat.test.wide[1] = n


fp = file.path('..', 'reports')
fn = 'Desiree.stat-omics-leaves.xlsx'
openxlsx::write.xlsx(stat.test.wide, file.path(fp, fn), asTable = TRUE)


stat.test.wide.leaves = stat.test.wide[, grep('Variable|PermP', colnames(stat.test.wide))]




```


# tubers

## data


```{r, tubers, echo=TRUE, warning=FALSE, message=FALSE}

fp = file.path('..', 'input')
fn = 'data_targeted.xlsx'

tubers = openxlsx::read.xlsx(xlsxFile = file.path(fp, fn),
                            sheet = 'metabolomics-tubers',
                            startRow = 1,
                            colNames = TRUE,
                            rowNames = FALSE,
                            detectDates = FALSE,
                            skipEmptyRows = TRUE,
                            skipEmptyCols = TRUE,
                            rows = NULL,
                            cols = NULL,
                            check.names = FALSE,
                            sep.names = ".",
                            namedRegion = NULL,
                            na.strings = "NA",
                            fillMergedCells = FALSE)
data.table::setDF(tubers)


```


```{r, DTC-tubers, echo=TRUE, warning=FALSE, message=FALSE}

stat.test = NULL


tubers = tubers[order(tubers$Stress, tubers$Tuber, tubers$Plant), ]
rownames(tubers) = paste(tubers[, 2], tubers[, 3], tubers[, 4], sep = '_')


DTC = tubers[, -grep('Plant|SampleID', colnames(tubers))]
DTC$Stress = factor(DTC$Stress, levels = c("C",   "D",   "H",   "HD",  "W"))
DTC = tidyr::gather(DTC, variable, measurement, colnames(DTC)[3]:colnames(DTC)[ncol(DTC)], factor_key=TRUE)
DTC = DTC[!is.na(DTC$measurement), ]
DTC = gdata::drop.levels(DTC)
DTC$variable = factor(DTC$variable)


```



## stat test

```{r, stat-tubers, echo=TRUE, warning=FALSE, message=FALSE}


en = levels(DTC$variable)[1]
eff = 'D'

stat.test = NULL


for (en in levels(DTC$variable)) {


  set.seed(seed)

  el = NULL
  el = levels(DTC$Stress)[-1]


  for (eff in el) {


    # cat(en, '\t', eff, '\n')






  mydata.long = NULL
  mydata.long = dplyr::as_tibble(data.table::data.table(DTC[DTC$variable == en &
                                                                   DTC$Stress %in% c('C', eff), ]))
  
  mydata.long = gdata::drop.levels(mydata.long)
  mytable = NULL
  
  
  model = MKinfer::perm.t.test(log2(measurement) ~ Stress ,
                                                   data=mydata.long,
                                                   alternative = "two.sided",
                                                   mu = 0,
                                                   paired = FALSE,
                                                   var.equal = FALSE,
                                                   conf.level = 0.95,
                                                   symmetric = TRUE, # a logical variable indicating whether to assume symmetry in the two-sided test. If TRUE then the symmetric permutation p value otherwise the equal-tail permutation p value is computed.
                                                   p.adjust.method = "none",
                                                   detailed = TRUE,
                                                   R = 999, set.seed(123456),
                                              simplify = TRUE)
  myvec = c(en, eff, model$method, model$p.value, model$perm.p.value)
  mytable = rbind(mytable, myvec)
  
  mytable = as.data.frame(mytable)
  colnames(mytable) = c('Variable', 'Treatment', 'Method', 'WelchP', 'PermP')
  
  stat.test = rbind(stat.test, mytable)



  }



} 



```


## wide format

```{r, wide-tubers, echo=TRUE, warning=FALSE, message=FALSE}


stat.test$Treatment = paste0(stat.test$Treatment, 
                                 formatC(28, width=2, flag="0"))

stat.test$WelchP = format(round(as.numeric(stat.test$WelchP), 3), nsmall = 3)
stat.test$PermP = format(round(as.numeric(stat.test$PermP), 3), nsmall = 3)


stat.test.wide2 = tidyr::spread(stat.test[, c(1:2, 5)], Treatment, PermP)
stat.test.wide1 = tidyr::spread(stat.test[, c(1:2, 4)], Treatment, WelchP)

colnames(stat.test.wide2)[2:ncol(stat.test.wide2)] = paste(colnames(stat.test.wide2)[2:ncol(stat.test.wide2)], 
                                                                   'PermP', sep = ' | ')
colnames(stat.test.wide1)[2:ncol(stat.test.wide1)] = paste(colnames(stat.test.wide1)[2:ncol(stat.test.wide1)], 
                                                                   'WelchP', sep = ' | ')


stat.test.wide = merge(stat.test.wide1, stat.test.wide2, by = c('Variable'), all = TRUE)
ind = match(sort(colnames(stat.test.wide)[2:ncol(stat.test.wide)]), colnames(stat.test.wide)[2:ncol(stat.test.wide)])
stat.test.wide = stat.test.wide[, c(1, ind+1)]

```

## write

```{r, res-tubers, echo=TRUE, warning=FALSE, message=FALSE}


fp = file.path('..', 'reports')
fn = 'Desiree.stat-omics-tubers.xlsx'
openxlsx::write.xlsx(stat.test.wide, file.path(fp, fn), asTable = TRUE)



```

# write res

```{r, res-write, echo=TRUE, warning=FALSE, message=FALSE}


fp = file.path('..', 'output')
fn = 'targeted_stat.txt'
stat.test.wide = stat.test.wide[, grep('Variable|PermP', colnames(stat.test.wide))]
stat.test.wide = merge(stat.test.wide.leaves, stat.test.wide, by = 'Variable', all.x = TRUE, all.y = TRUE)

data.table::fwrite(stat.test.wide, file.path(fp, fn), sep = '\t')


```




# session Info

```{r, echo=TRUE, warning=FALSE, message=FALSE}

sessionInfo()

```
