==> _p_ADAPTOmics/_I_Desiree/_S_multiOmics/_A_multiOmicsStat-R/scripts/04.3_plot.Rmd <==
---
title: "04.3_plot"
author: "zagor"
date: "`r Sys.Date()`"
output:
  html_document:
    fig_caption: yes
    self_contained: yes
    fig_width: 9
    fig_height: 6
    toc: true
    toc_float:
      toc_collapsed: false
    toc_depth: 5
    number_sections: yes
    theme: flatly
    highlight: tango
  pdf_document:
    toc: yes
    toc_depth: '5'
  word_document:
    toc: yes
    toc_depth: '5'
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(#dev = c('pdf', 'png'),  # this embeds pdf and crates scrolable blocks
                      dev = c('png'), 
                      fig.align = 'center', 
                      fig.height = 6, 
                      fig.width = 9 ,
                      warning = FALSE, message = FALSE
                      )

```



```{r, rm, echo=TRUE, results='hide', warning=FALSE, message=FALSE}

rm(list = ls(all = TRUE))

gc()
gc()

seed = 123456
set.seed(seed)

```


# libraries

```{r, libraries, echo=TRUE, warning=FALSE, message=FALSE}





library(magrittr)

library(pheatmap)

library(RColorBrewer)

library(ComplexHeatmap)



```


# palette

```{r, echo=TRUE, warning=FALSE, message=FALSE}

# col = circlize::colorRamp2(c(-3.5,-1,0,1,3.5), c("darkblue","skyblue","white", "brown1", "darkred"))
# col = circlize::colorRamp2(c(-3.5,-1,0,1,3.5), c(brewer.pal(9, 'Blues')[8],"skyblue","white", "brown1", rev(brewer.pal(9, 'Reds'))[2]))
col = circlize::colorRamp2(c(-3.0, 0, 3.0), c("darkblue","white", "darkred"))

```

# data

```{r, echo=TRUE, warning=FALSE, message=FALSE}


fp = file.path('..', 'reports')
fn = 'Desiree.targeted_logFC-sig.xlsx'

plotMe = openxlsx::read.xlsx(xlsxFile = file.path(fp, fn),
                            sheet = 1,
                            startRow = 1,
                            colNames = TRUE,
                            rowNames = FALSE,
                            detectDates = FALSE,
                            skipEmptyRows = TRUE,
                            skipEmptyCols = TRUE,
                            rows = NULL,
                            cols = NULL,
                            check.names = FALSE,
                            sep.names = ".",
                            namedRegion = NULL,
                            na.strings = "NA",
                            fillMergedCells = FALSE)

colnames(plotMe) = gsub('\\||logFC|\\.|_', '', colnames(plotMe))
rownames(plotMe) = plotMe$molecule


mygroup = unique(plotMe$group)



```

# H, HD, D

```{r, echo=TRUE, warning=FALSE, message=FALSE}

plotMe.H = plotMe[, setdiff(grep('^H|^D', colnames(plotMe)), grep('28', colnames(plotMe)))]
rs = rowSums(is.na(plotMe.H))/ncol(plotMe.H)
plotMe.H = plotMe.H[-which(rs == 1), ]



for (i in mygroup) {
  
  sel = plotMe$molecule[plotMe$group == i]
  tmp = plotMe.H[rownames(plotMe.H) %in% sel, ]
  
  row_colors = plotMe[match(rownames(tmp), plotMe$molecule), ]$colourCode

  pdf(file.path('..', 'other', 'plots', paste0('F5_', i, '.pdf')), width=4,height=4)
  ht = Heatmap(as.matrix(tmp), 
          col = col,
          na_col = 'white',
          column_title = i, 
          show_heatmap_legend = FALSE,
          heatmap_legend_param = list(title = "legend", at = c(-3.0, 0, 3.0)), # seq(-3.5, 3.5, length.out = length(col) + 1)), 
          row_names_gp = grid::gpar(fontsize = 9, col = row_colors, fontface = 'bold'),
          column_names_gp = grid::gpar(fontsize = 9, fontface = 'bold'),
          cluster_columns = FALSE, cluster_rows = FALSE,
          rect_gp = gpar(col = "grey80", lwd = 1),
          width = ncol(tmp)*unit(5, "mm"),
          height = unit(55, "mm") # nrow(tmp)*unit(5, "mm")
          )
  draw(ht)
  dev.off()

}



```

# W, D

```{r, echo=TRUE, warning=FALSE, message=FALSE}


plotMe.W = plotMe[, setdiff(grep('^W|^D', colnames(plotMe)), grep('28', colnames(plotMe)))]
plotMe.W = plotMe.W[, c(3,4,2)]
rs = rowSums(is.na(plotMe.W))/ncol(plotMe.W)
plotMe.W = plotMe.W[-which(rs == 1), ]

for (i in mygroup) {
  
  sel = plotMe$molecule[plotMe$group == i]
  tmp = plotMe.W[rownames(plotMe.W) %in% sel, ]
  
  row_colors = plotMe[match(rownames(tmp), plotMe$molecule), ]$colourCode

  pdf(file.path('..', 'other', 'plots', paste0('F6_', i, '.pdf')), width=4,height=4)
  ht = Heatmap(as.matrix(tmp), 
          col = col,
          na_col = 'white',
          column_title = i, 
          show_heatmap_legend = FALSE,
          heatmap_legend_param = list(title = "legend", at = c(-3.5,0,3.5)), # seq(-3.5, 3.5, length.out = length(col) + 1)), 
          row_names_gp = grid::gpar(fontsize = 9, col = row_colors, fontface = 'bold'),
          column_names_gp = grid::gpar(fontsize = 9, fontface = 'bold'),
          cluster_columns = FALSE, cluster_rows = FALSE,
          rect_gp = gpar(col = "grey80", lwd = 1),
          width = ncol(tmp)*unit(5, "mm"),
          height = unit(55, "mm") # nrow(tmp)*unit(5, "mm")
          )
  draw(ht)
  dev.off()

}

```





# session Info

```{r, echo=TRUE, warning=FALSE, message=FALSE}

sessionInfo()

```
